### 2026-03-26 7b7c3c4
Fix bib and add country to affiliations
