+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @misc{boost,
-  author = {Boost Authors},
+  author = {{Boost Authors}},
   title = {Boost C++ Libraries Version},
   year = {2025},
   publisher = {GitHub},
@@ -20,7 +20,7 @@
 }
 
 @misc{ibm_libdfp,
-  author = {IBM Corporation},
+  author = {{IBM Corporation}},
   title = {Decimal Floating Point C Library},
   year = {2010},
   publisher = {GitHub},
@@ -29,7 +29,7 @@
 }
 
 @misc{intel_dfp,
-  author = {Intel Corporation},
+  author = {{Intel Corporation}},
   title = {Intel(R) Decimal Floating-Point Math Library v2.3},
   year = {2023},
   publisher = {netlib},
