### 2026-03-26 #10302 @mborland
> 👋 [@mborland](https://github.com/mborland) - thanks for your submission. Please add countries to the 2 affiliations. Please also add an extra set of {}s around the author data for the first three references, as these are not human names that bibtex should parse but strings that it should just use. You can then use the command `@editorialbot generate pdf` command to regenerate the pdf. An editorialbot command needs to be the first entry in a new comment.

These two items have been fixed.
