+++ test/test_cbrt_from_math.cpp
@@ -1,0 +1,259 @@
+// Copyright 2026 Christopher Kormanyos
+// Distributed under the Boost Software License, Version 1.0.
+// https://www.boost.org/LICENSE_1_0.txt
+
+#include "testing_config.hpp"
+
+#include <chrono>
+#include <iomanip>
+#include <iostream>
+#include <limits>
+#include <random>
+#include <type_traits>
+
+#include <boost/decimal.hpp>
+#include <boost/math/policies/policy.hpp>
+#include <boost/math/special_functions/cbrt.hpp>
+#include <boost/multiprecision/cpp_dec_float.hpp>
+
+#if defined(__clang__)
+#  pragma clang diagnostic push
+#  pragma clang diagnostic ignored "-Wfloat-equal"
+#elif defined(__GNUC__)
+#  pragma GCC diagnostic push
+#  pragma GCC diagnostic ignored "-Wfloat-equal"
+#endif
+
+#include <boost/core/lightweight_test.hpp>
+
+namespace boost { namespace math { namespace policies {
+
+// Specialization of the precision structure.
+template<typename ThisPolicy>
+struct precision<boost::decimal::decimal128_t, ThisPolicy>
+{
+  using local_decimal_type = boost::decimal::decimal128_t;
+
+  using precision_type = typename ThisPolicy::precision_type;
+
+  using local_digits_2 = digits2<((static_cast<std::intmax_t>(std::numeric_limits<local_decimal_type>::digits10) + INTMAX_C(1)) * INTMAX_C(1000)) / INTMAX_C(301)>; // NOLINT(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
+
+  using type =
+    typename std::conditional<((local_digits_2::value <= precision_type::value) || (precision_type::value <= 0)),
+                                local_digits_2,
+                                precision_type>::type;
+};
+
+} } } // namespace boost::math::policies
+
+
+namespace local
+{
+  template<typename IntegralTimePointType,
+           typename ClockType = std::chrono::high_resolution_clock>
+  auto time_point() noexcept -> IntegralTimePointType
+  {
+    using local_integral_time_point_type = IntegralTimePointType;
+    using local_clock_type               = ClockType;
+
+    const auto current_now =
+      static_cast<std::uintmax_t>
+      (
+        std::chrono::duration_cast<std::chrono::nanoseconds>
+        (
+          local_clock_type::now().time_since_epoch()
+        ).count()
+      );
+
+    return static_cast<local_integral_time_point_type>(current_now);
+  }
+
+  template<typename NumericType>
+  auto is_close_fraction(const NumericType& a,
+                         const NumericType& b,
+                         const NumericType& tol) noexcept -> bool
+  {
+    using std::fabs;
+
+    auto result_is_ok = bool { };
+
+    if(b == static_cast<NumericType>(0))
+    {
+      result_is_ok = (fabs(a - b) < tol); // LCOV_EXCL_LINE
+    }
+    else
+    {
+      const auto delta = fabs(1 - (a / b));
+
+      result_is_ok = (delta < tol);
+    }
+
+    return result_is_ok;
+  }
+
+  template<typename ExtendedFloatType>
+  auto get_random_float_string(std::string* p_str_flt = nullptr, const bool do_seed_rnd_gens = false) -> ExtendedFloatType; // NOLINT(google-runtime-references)
+
+  template<typename ExtendedFloatType>
+  auto get_random_float_string(std::string* p_str_flt, const bool do_seed_rnd_gens) -> ExtendedFloatType // NOLINT(google-runtime-references)
+  {
[…]
