### 2026-03-29 d1d4299
Handle a compiler warning
