### 2026-03-29 c19bbbe
Handle more compiler warnings
