### 2026-03-29 4715740
Handle an unused variable in new test
