### 2026-03-29 8f570bd
Disable selected warns in new test
