+++ test/test_cbrt_from_math.cpp
@@ -18,11 +18,15 @@
 #  pragma clang diagnostic ignored "-Wfloat-equal"
 #  pragma clang diagnostic push
 #  pragma clang diagnostic ignored "-Wold-style-cast"
+#  pragma clang diagnostic push
+#  pragma clang diagnostic ignored "-Wundef"
 #elif defined(__GNUC__)
 #  pragma GCC diagnostic push
 #  pragma GCC diagnostic ignored "-Wfloat-equal"
 #  pragma GCC diagnostic push
 #  pragma GCC diagnostic ignored "-Wold-style-cast"
+#  pragma GCC diagnostic push
+#  pragma GCC diagnostic ignored "-Wuseless-cast"
 #endif
 
 #include <boost/math/policies/policy.hpp>
@@ -257,7 +261,9 @@ auto main() -> int
 #if defined(__clang__)
 #  pragma clang diagnostic pop
 #  pragma clang diagnostic pop
+#  pragma clang diagnostic pop
 #elif defined(__GNUC__)
+#  pragma GCC diagnostic pop
 #  pragma GCC diagnostic pop
 #  pragma GCC diagnostic pop
 #endif
