### 2026-03-30 0d3ffc7
Handle another warning disable
