+++ test/test_cbrt_from_math.cpp
@@ -48,11 +48,15 @@
 #  pragma clang diagnostic ignored "-Wold-style-cast"
 #  pragma clang diagnostic push
 #  pragma clang diagnostic ignored "-Wundef"
+#  pragma clang diagnostic push
+#  pragma clang diagnostic ignored "-Wunused-parameter"
 #elif defined(__GNUC__)
 #  pragma GCC diagnostic push
 #  pragma GCC diagnostic ignored "-Wfloat-equal"
 #  pragma GCC diagnostic push
 #  pragma GCC diagnostic ignored "-Wold-style-cast"
+#  pragma GCC diagnostic push
+#  pragma GCC diagnostic ignored "-Wunused-parameter"
 #  pragma GCC diagnostic push
 #  pragma GCC diagnostic ignored "-Wuseless-cast"
 #endif
@@ -293,7 +297,9 @@ auto main() -> int
 #  pragma clang diagnostic pop
 #  pragma clang diagnostic pop
 #  pragma clang diagnostic pop
+#  pragma clang diagnostic pop
 #elif defined(__GNUC__)
+#  pragma GCC diagnostic pop
 #  pragma GCC diagnostic pop
 #  pragma GCC diagnostic pop
 #  pragma GCC diagnostic pop
