### 2026-03-31 58ce113
Fix or ignore new GCC warnings
