### 2026-03-31 51f3cdd
Clang doesn't know useless-cast
