### 2026-03-31 54b6f3b
Add basic configuration
