### 2026-04-01 da0327b
Bulk annotation of functions
