+++ include/boost/decimal/detail/add_impl.hpp
@@ -29,7 +29,7 @@ namespace detail {
 #endif
 
 template <typename ReturnType, typename T>
-constexpr auto add_impl(const T& lhs, const T& rhs) noexcept -> ReturnType
+BOOST_DECIMAL_CUDA_CONSTEXPR auto add_impl(const T& lhs, const T& rhs) noexcept -> ReturnType
 {
     // Each of the significands is maximally 23 bits.
     // Rather than doing division to get proper alignment we will promote to 64 bits
