+++ include/boost/decimal/detail/fenv_rounding.hpp
@@ -44,7 +44,7 @@ struct divmod10_result
 #endif
 
 template <typename T, std::enable_if_t<std::is_integral<T>::value, bool> = true>
-constexpr auto divmod(T dividend, T divisor) noexcept -> divmod_result<T>
+BOOST_DECIMAL_CUDA_CONSTEXPR auto divmod(T dividend, T divisor) noexcept -> divmod_result<T>
 {
     // Compilers usually can lump these together
     const auto q {static_cast<T>(dividend / divisor)};
@@ -53,7 +53,7 @@ constexpr auto divmod(T dividend, T divisor) noexcept -> divmod_result<T>
 }
 
 template <typename T, std::enable_if_t<!std::is_integral<T>::value, bool> = true>
-constexpr auto divmod(T dividend, T divisor) noexcept -> divmod_result<T>
+BOOST_DECIMAL_CUDA_CONSTEXPR auto divmod(T dividend, T divisor) noexcept -> divmod_result<T>
 {
     T q {dividend / divisor};
     T r {dividend - q * divisor};
@@ -71,26 +71,26 @@ constexpr auto divmod(const int128::uint128_t dividend, const int128::uint128_t 
 
 #endif
 
-constexpr auto divmod(const u256& lhs, const u256& rhs) noexcept
+BOOST_DECIMAL_CUDA_CONSTEXPR auto divmod(const u256& lhs, const u256& rhs) noexcept
 {
     return div_mod(lhs, rhs);
 }
 
 template <typename T>
-constexpr auto divmod10(const T dividend) noexcept -> divmod10_result<T>
+BOOST_DECIMAL_CUDA_CONSTEXPR auto divmod10(const T dividend) noexcept -> divmod10_result<T>
 {
     const auto q {static_cast<T>(dividend / 10U)};
     const auto r {static_cast<T>(dividend - q * 10U)};
     return {q, static_cast<std::uint64_t>(r)};
 }
 
-constexpr auto divmod10(const u256& lhs) noexcept
+BOOST_DECIMAL_CUDA_CONSTEXPR auto divmod10(const u256& lhs) noexcept
 {
     constexpr int128::uint128_t ten {10U};
     return div_mod(lhs, ten);
 }
 
-constexpr auto divmod10(const int128::uint128_t lhs) noexcept -> divmod10_result<int128::uint128_t>
+BOOST_DECIMAL_CUDA_CONSTEXPR auto divmod10(const int128::uint128_t lhs) noexcept -> divmod10_result<int128::uint128_t>
 {
     constexpr std::uint32_t ten {10U};
     int128::uint128_t q {};
@@ -100,7 +100,7 @@ constexpr auto divmod10(const int128::uint128_t lhs) noexcept -> divmod10_result
 }
 
 template <typename TargetType, typename T>
-constexpr auto fenv_round_impl(T& val, const bool is_neg, const bool sticky, const rounding_mode round = _boost_decimal_global_rounding_mode) noexcept -> int
+BOOST_DECIMAL_CUDA_CONSTEXPR auto fenv_round_impl(T& val, const bool is_neg, const bool sticky, const rounding_mode round = _boost_decimal_global_rounding_mode) noexcept -> int
 {
     using significand_type = std::conditional_t<decimal_val_v<TargetType> >= 128, int128::uint128_t, std::int64_t>;
 
@@ -163,7 +163,7 @@ constexpr auto fenv_round_impl(T& val, const bool is_neg, const bool sticky, con
 #ifdef BOOST_DECIMAL_NO_CONSTEVAL_DETECTION
 
 template <typename TargetType, typename T, std::enable_if_t<is_integral_v<T>, bool> = true>
[…]

[…]
