### 2026-04-01 27d7df4
Add basic operator tests and CUDA specific Jamfile
