### 2026-04-01 dba7261
Fix warnings and errors
