+++ test/test_cuda_decimal32_t_add.cu
@@ -3,8 +3,6 @@
 //  Boost Software License, Version 1.0. (See accompanying file
 //  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#define BOOST_DECIMAL_ENABLE_CUDA
-
 #include <iostream>
 #include <vector>
 #include <random>
