### 2026-04-01 ce40048
Copy tables into functions when using CUDA
