+++ include/boost/decimal/detail/power_tables.hpp
@@ -20,6 +20,8 @@ namespace detail {
 
 namespace impl {
 
+#if !defined(__CUDACC__) || !defined(BOOST_DECIMAL_ENABLE_CUDA)
+
 BOOST_DECIMAL_INLINE_CONSTEXPR_VARIABLE std::uint32_t powers_of_10_u32[10] =
 {
     UINT32_C(1), UINT32_C(10), UINT32_C(100), UINT32_C(1000), UINT32_C(10000), UINT32_C(100000), UINT32_C(1000000),
@@ -206,6 +208,8 @@ BOOST_DECIMAL_INLINE_CONSTEXPR_VARIABLE u256 u256_pow_10[79] = {
     u256{UINT64_C(11735238523568814774), UINT64_C(16381442489372128114), UINT64_C(12246644529347313664), UINT64_C(0)},
 };
 
+#endif // !defined(__CUDACC__) || !defined(BOOST_DECIMAL_ENABLE_CUDA)
+
 } // namespace impl
 
 #if defined(__GNUC__) && __GNUC__ >= 7
@@ -216,19 +220,94 @@ BOOST_DECIMAL_INLINE_CONSTEXPR_VARIABLE u256 u256_pow_10[79] = {
 template <typename T>
 BOOST_DECIMAL_CUDA_CONSTEXPR auto pow10(T n) noexcept -> T
 {
+    #if defined(__CUDACC__) && defined(BOOST_DECIMAL_ENABLE_CUDA)
+
+    constexpr std::uint64_t powers_of_10[20] =
+    {
+        UINT64_C(1), UINT64_C(10), UINT64_C(100), UINT64_C(1000), UINT64_C(10000), UINT64_C(100000), UINT64_C(1000000),
+        UINT64_C(10000000), UINT64_C(100000000), UINT64_C(1000000000), UINT64_C(10000000000), UINT64_C(100000000000),
+        UINT64_C(1000000000000), UINT64_C(10000000000000), UINT64_C(100000000000000), UINT64_C(1000000000000000),
+        UINT64_C(10000000000000000), UINT64_C(100000000000000000), UINT64_C(1000000000000000000),
+        UINT64_C(10000000000000000000)
+    };
+    return static_cast<T>(powers_of_10[static_cast<std::size_t>(n)]);
+
+    #else
+
     return static_cast<T>(impl::powers_of_10[static_cast<std::size_t>(n)]);
+
+    #endif
 }
 
 template <>
 BOOST_DECIMAL_CUDA_CONSTEXPR auto pow10(const boost::int128::uint128_t n) noexcept -> boost::int128::uint128_t
 {
+    #if defined(__CUDACC__) && defined(BOOST_DECIMAL_ENABLE_CUDA)
+
+    constexpr boost::int128::uint128_t BOOST_DECIMAL_DETAIL_INT128_pow10[39] =
+    {
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(1),
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(10),
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(100),
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(1000),
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(10000),
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(100000),
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(1000000),
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(10000000),
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(100000000),
+        BOOST_DECIMAL_DETAIL_INT128_UINT128_C(1000000000),
[…]
