### 2026-04-01 4861ddf
Add CUDA safe max function
