### 2026-04-01 7556105
Fixup 256 bit table reads and functions
