+++ include/boost/decimal/detail/power_tables.hpp
@@ -320,14 +320,108 @@ constexpr auto pow10(detail::builtin_uint128_t n) noexcept -> detail::builtin_ui
 
 #endif
 
-BOOST_DECIMAL_CUDA_CONSTEXPR auto pow10(const u256& n) noexcept -> u256
-{
-    return impl::u256_pow_10[static_cast<std::uint64_t>(n)];
-}
-
 BOOST_DECIMAL_CUDA_CONSTEXPR auto pow10_256(const std::size_t n) noexcept -> u256
 {
+    #if defined(__CUDACC__) && defined(BOOST_DECIMAL_ENABLE_CUDA)
+
+    constexpr u256 u256_pow_10[79] = {
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(1)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(10)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(100)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(1000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(10000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(100000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(1000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(10000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(100000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(1000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(10000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(100000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(1000000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(10000000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(100000000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(1000000000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(10000000000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(100000000000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(1000000000000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(0), UINT64_C(10000000000000000000)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(5), UINT64_C(7766279631452241920)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(54), UINT64_C(3875820019684212736)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(542), UINT64_C(1864712049423024128)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(5421), UINT64_C(200376420520689664)},
+        u256{UINT64_C(0), UINT64_C(0), UINT64_C(54210), UINT64_C(2003764205206896640)},
[…]
