### 2026-04-01 b4d0eaa
Call functions instead of tables
