### 2026-04-01 ea508bf
More syntax and tests
