### 2026-04-01 9b2f97c
Add markers to std::numeric_limits usages
