### 2026-04-01 5463cda
Add CUDA CMake workflow to test compilation at least
