### 2026-04-01 3190f03
Add CUDA annotations for basic decimal64_t functionality
