+++ include/boost/decimal/decimal64_t.hpp
@@ -128,22 +128,22 @@ private:
     std::uint64_t bits_ {};
 
     // Returns the un-biased (quantum) exponent
-    constexpr auto unbiased_exponent() const noexcept -> exponent_type;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto unbiased_exponent() const noexcept -> exponent_type;
 
     // Returns the biased exponent
-    constexpr auto biased_exponent() const noexcept -> biased_exponent_type;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto biased_exponent() const noexcept -> biased_exponent_type;
 
     // Allows direct editing of the exp
     template <typename T>
-    constexpr auto edit_exponent(T exp) noexcept
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto edit_exponent(T exp) noexcept
         BOOST_DECIMAL_REQUIRES_RETURN(detail::is_integral_v, T, void);
 
     // Returns the significand complete with the bits implied from the combination field
-    constexpr auto full_significand() const noexcept -> significand_type;
-    constexpr auto isneg() const noexcept -> bool;
-    constexpr auto edit_sign(bool sign) noexcept -> void;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto full_significand() const noexcept -> significand_type;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto isneg() const noexcept -> bool;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto edit_sign(bool sign) noexcept -> void;
 
-    constexpr auto to_components() const noexcept -> detail::decimal64_t_components;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto to_components() const noexcept -> detail::decimal64_t_components;
 
     // Attempts conversion to integral type:
     // If this is nan sets errno to EINVAL and returns 0
@@ -160,8 +160,8 @@ private:
     friend constexpr auto to_decimal(Decimal val) noexcept -> TargetType;
 
     // Debug bit pattern
-    friend constexpr auto from_bits(std::uint64_t bits) noexcept -> decimal64_t;
-    friend constexpr auto to_bits(decimal64_t rhs) noexcept -> std::uint64_t;
+    friend BOOST_DECIMAL_CUDA_CONSTEXPR auto from_bits(std::uint64_t bits) noexcept -> decimal64_t;
+    friend BOOST_DECIMAL_CUDA_CONSTEXPR auto to_bits(decimal64_t rhs) noexcept -> std::uint64_t;
 
     // Equality template between any integer type and decimal64_t
     template <BOOST_DECIMAL_DECIMAL_FLOATING_TYPE Decimal, BOOST_DECIMAL_INTEGRAL Integer>
@@ -183,7 +183,7 @@ private:
         -> std::enable_if_t<(detail::is_decimal_floating_point_v<Decimal1> &&
                              detail::is_decimal_floating_point_v<Decimal2>), bool>;
 
-    friend constexpr auto d64_div_impl(decimal64_t lhs, decimal64_t rhs, decimal64_t& q, decimal64_t& r) noexcept -> void;
+    friend BOOST_DECIMAL_CUDA_CONSTEXPR auto d64_div_impl(decimal64_t lhs, decimal64_t rhs, decimal64_t& q, decimal64_t& r) noexcept -> void;
 
     template <typename T>
[…]
