### 2026-04-01 e2ee3c5
Add testing of basic decimal64_t functionality on device
