### 2026-04-01 bfa7219
Avoid usage of builtin 128-bit integers on device
