### 2026-04-01 a651a67
Remove access to global for rounding
