### 2026-04-02 #10345 @editorialbot
[REVIEW]: Boost.Decimal: A C++14 Library for Decimal Floating Point Arithmetic
**Submitting author:** @mborland (<a href="http://orcid.org/0009-0005-8183-4254">Matt Borland</a>)
**Repository:** https://github.com/boostorg/decimal
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v6.0.1
