### 2026-04-02 960e949
Repair trivial casting warning
