### 2026-04-02 0bcb79b
Increase the size of the stack from 1024 to 2048
