+++ test/test_cuda_decimal64_t_add.cu
@@ -46,6 +46,7 @@ int main(void)
 
     watch w;
 
+    cudaDeviceSetLimit(cudaLimitStackSize, 2048);
     cuda_test<<<blocksPerGrid, threadsPerBlock>>>(input_vector.get(), output_vector.get(), numElements);
     cudaDeviceSynchronize();
 
