### 2026-04-03 c5826d7
Add markers for CUDA support to decimal128_t
