+++ include/boost/decimal/decimal128_t.hpp
@@ -122,18 +122,18 @@ private:
 
     #endif
 
-    friend constexpr auto from_bits(int128::uint128_t rhs) noexcept -> decimal128_t;
+    friend BOOST_DECIMAL_CUDA_CONSTEXPR auto from_bits(int128::uint128_t rhs) noexcept -> decimal128_t;
 
-    constexpr auto unbiased_exponent() const noexcept -> std::uint64_t;
-    constexpr auto biased_exponent() const noexcept -> std::int32_t;
-    constexpr auto full_significand() const noexcept -> int128::uint128_t;
-    constexpr auto isneg() const noexcept -> bool;
-    constexpr auto to_components() const noexcept -> detail::decimal128_t_components;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto unbiased_exponent() const noexcept -> std::uint64_t;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto biased_exponent() const noexcept -> std::int32_t;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto full_significand() const noexcept -> int128::uint128_t;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto isneg() const noexcept -> bool;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto to_components() const noexcept -> detail::decimal128_t_components;
 
     // Allows direct editing of the exp
     template <typename T, std::enable_if_t<detail::is_integral_v<T>, bool> = true>
-    constexpr auto edit_exponent(T exp) noexcept -> void;
-    constexpr auto edit_sign(bool sign) noexcept -> void;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto edit_exponent(T exp) noexcept -> void;
+    BOOST_DECIMAL_CUDA_CONSTEXPR auto edit_sign(bool sign) noexcept -> void;
 
     // Attempts conversion to integral type:
     // If this is nan sets errno to EINVAL and returns 0
@@ -172,7 +172,7 @@ private:
         -> std::enable_if_t<(detail::is_decimal_floating_point_v<Decimal1> &&
                              detail::is_decimal_floating_point_v<Decimal2>), bool>;
 
-    friend constexpr auto d128_div_impl(const decimal128_t& lhs, const decimal128_t& rhs, decimal128_t& q, decimal128_t& r) noexcept -> void;
+    friend BOOST_DECIMAL_CUDA_CONSTEXPR auto d128_div_impl(const decimal128_t& lhs, const decimal128_t& rhs, decimal128_t& q, decimal128_t& r) noexcept -> void;
 
     template <typename T>
     friend constexpr auto ilogb(T d) noexcept
@@ -182,7 +182,7 @@ private:
     friend constexpr auto logb(T num) noexcept
         BOOST_DECIMAL_REQUIRES(detail::is_decimal_floating_point_v, T);
 
-    friend constexpr auto not_finite(decimal128_t rhs) noexcept -> bool;
+    friend BOOST_DECIMAL_CUDA_CONSTEXPR auto not_finite(decimal128_t rhs) noexcept -> bool;
 
     friend constexpr auto to_bid_d128(decimal128_t val) noexcept -> int128::uint128_t;
 
@@ -223,7 +223,7 @@ private:
     friend constexpr auto detail::write_payload(typename TargetDecimalType::significand_type payload_value)
         BOOST_DECIMAL_REQUIRES(detail::is_ieee_type_v, TargetDecimalType);
 
[…]

[…]
