### 2026-04-03 2e6d36f
Add decimal128_t cuda testing
