### 2026-04-03 e424cd7
Replace use of global variable with template variable
