### 2026-04-03 1329ea1
Add workaround for NVCC
