+++ include/boost/decimal/decimal128_t.hpp
@@ -709,7 +709,10 @@ BOOST_DECIMAL_CUDA_CONSTEXPR decimal128_t::decimal128_t(T1 coeff, T2 exp, const 
     auto biased_exp {static_cast<int>(exp + detail::bias_v<decimal128_t>)};
     BOOST_DECIMAL_IF_CONSTEXPR (sizeof(T1) >= sizeof(significand_type))
     {
-        if (coeff > detail::max_significand_v<decimal128_t> || biased_exp < -(detail::precision_v<decimal128_t> - 1))
+        // NVCC can't access uint128_t variables since it's a non-trivial struct.
+        // We need to call the related function and store the result instead.
+        constexpr auto max_sig {detail::impl::max_significand_v<decimal128_t>()};
+        if (coeff > max_sig || biased_exp < -(detail::precision_v<decimal128_t> - 1))
         {
             coeff_digits = detail::coefficient_rounding<decimal128_t>(coeff, exp, biased_exp, is_negative, detail::num_digits(coeff));
         }
