### 2026-04-03 a4a9203
Increase stack size
