+++ test/test_cuda_decimal128_t_add.cu
@@ -46,7 +46,7 @@ int main(void)
 
     watch w;
 
-    cudaDeviceSetLimit(cudaLimitStackSize, 2048);
+    cudaDeviceSetLimit(cudaLimitStackSize, 4096);
     cuda_test<<<blocksPerGrid, threadsPerBlock>>>(input_vector.get(), output_vector.get(), numElements);
     cudaDeviceSynchronize();
 
