### 2026-04-03 6c75014
Further increase stack size
