+++ test/test_cuda_decimal128_t_div.cu
@@ -49,7 +49,7 @@ int main(void)
 
     watch w;
 
-    cudaDeviceSetLimit(cudaLimitStackSize, 4096);
+    cudaDeviceSetLimit(cudaLimitStackSize, 8192);
     cuda_test<<<blocksPerGrid, threadsPerBlock>>>(input_vector.get(), input_vector2.get(), output_vector.get(), numElements);
     cudaDeviceSynchronize();
 
