### 2026-04-03 4846a0a
Fix erroneous casts to unrepresentable values
