### 2026-04-03 7a887b7
Fix erroneous casts to unrepresentable values
