### 2026-04-03 8d222dc
Bracket result instead of finding exact value
