### 2026-04-03 e3baf71
Use decimal one instead of integer literal 1
