### 2026-04-03 32d6aa0
Use decimal one instead of integer literal 1
