+++ include/boost/decimal/detail/to_integral.hpp
@@ -37,6 +37,7 @@ constexpr auto to_integral(Decimal val) noexcept
 
     constexpr Decimal max_target_type { (std::numeric_limits<TargetType>::max)() };
     constexpr Decimal min_target_type { (std::numeric_limits<TargetType>::min)() };
+    constexpr Decimal one {1};
 
     if (isnan(val))
     {
@@ -50,7 +51,7 @@ constexpr auto to_integral(Decimal val) noexcept
     }
 
     // Anything in [0, 1) should be flushed to 0
-    if (abs(val) < 1)
+    if (abs(val) < one)
     {
         return static_cast<TargetType>(0);
     }
@@ -104,6 +105,7 @@ constexpr auto to_integral_128(Decimal val) noexcept
 {
     constexpr Decimal max_target_type { (std::numeric_limits<TargetType>::max)() };
     constexpr Decimal min_target_type { (std::numeric_limits<TargetType>::min)() };
+    constexpr Decimal one {1};
 
     if (isnan(val))
     {
@@ -118,7 +120,7 @@ constexpr auto to_integral_128(Decimal val) noexcept
     }
 
     // Anything in [0, 1) should be flushed to 0
-    if (abs(val) < 1)
+    if (abs(val) < one)
     {
         return static_cast<TargetType>(0);
     }
