### 2026-04-03 8df9069
Make decimal32_t/decimal64_t/decimal128_t similar to regular int/double with no default initialization
closes #1364
