### 2026-04-03 9f833c9
Attempt to make charconv available on device
