### 2026-04-03 cf2f3ed
Move uchar values table into function
