### 2026-04-03 458f41e
Disable use of CLZ intrinsics on device
