### 2026-04-03 67808d3
Move CLZ table into function when on device
