### 2026-04-06 c44572c
Workaround for clang < 9 not being able to parse res string
