### 2026-04-06 2736d0b
Reduce timing of Clang 15-20 runs
