### 2026-04-06 0eed1cd
Reduce length of UBSAN runs
