+++ include/boost/decimal/decimal32_t.hpp
@@ -276,7 +276,7 @@ public:
     #else
     template <typename Decimal, std::enable_if_t<detail::is_decimal_floating_point_v<Decimal>, bool> = true>
     #endif
-    explicit constexpr decimal32_t(Decimal val) noexcept;
+    explicit BOOST_DECIMAL_CUDA_CONSTEXPR decimal32_t(Decimal val) noexcept;
 
     // 3.2.2.3 Conversion from integral type
     #ifdef BOOST_DECIMAL_HAS_CONCEPTS
@@ -1704,7 +1704,7 @@ template <BOOST_DECIMAL_DECIMAL_FLOATING_TYPE Decimal>
 #else
 template <typename Decimal, std::enable_if_t<detail::is_decimal_floating_point_v<Decimal>, bool>>
 #endif
-constexpr decimal32_t::decimal32_t(const Decimal val) noexcept
+BOOST_DECIMAL_CUDA_CONSTEXPR decimal32_t::decimal32_t(const Decimal val) noexcept
 {
     *this = to_decimal<decimal32_t, Decimal>(val);
 }
