+++ include/boost/decimal/decimal128_t.hpp
@@ -269,7 +269,7 @@ public:
     #else
     template <typename Decimal, std::enable_if_t<detail::is_decimal_floating_point_v<Decimal>, bool> = true>
     #endif
-    explicit constexpr decimal128_t(Decimal val) noexcept;
+    explicit BOOST_DECIMAL_CUDA_CONSTEXPR decimal128_t(Decimal val) noexcept;
 
     #ifdef BOOST_DECIMAL_HAS_CONCEPTS
     template <BOOST_DECIMAL_INTEGRAL Integer>
@@ -1057,7 +1057,7 @@ template <BOOST_DECIMAL_DECIMAL_FLOATING_TYPE Decimal>
 #else
 template <typename Decimal, std::enable_if_t<detail::is_decimal_floating_point_v<Decimal>, bool>>
 #endif
-constexpr decimal128_t::decimal128_t(const Decimal val) noexcept
+BOOST_DECIMAL_CUDA_CONSTEXPR decimal128_t::decimal128_t(const Decimal val) noexcept
 {
     *this = to_decimal<decimal128_t>(val);
 }
