+++ include/boost/decimal/decimal64_t.hpp
@@ -288,7 +288,7 @@ public:
     #else
     template <typename Decimal, std::enable_if_t<detail::is_decimal_floating_point_v<Decimal>, bool> = true>
     #endif
-    explicit constexpr decimal64_t(Decimal val) noexcept;
+    explicit BOOST_DECIMAL_CUDA_CONSTEXPR decimal64_t(Decimal val) noexcept;
 
     // 3.2.3.3 Conversion from integral type
     #ifdef BOOST_DECIMAL_HAS_CONCEPTS
@@ -980,7 +980,7 @@ template <BOOST_DECIMAL_DECIMAL_FLOATING_TYPE Decimal>
 #else
 template <typename Decimal, std::enable_if_t<detail::is_decimal_floating_point_v<Decimal>, bool>>
 #endif
-constexpr decimal64_t::decimal64_t(const Decimal val) noexcept
+BOOST_DECIMAL_CUDA_CONSTEXPR decimal64_t::decimal64_t(const Decimal val) noexcept
 {
     *this = to_decimal<decimal64_t>(val);
 }
