### 2026-04-09 7cbc0a9
Add testing of CUDA mixed decimal arithmetic
