+++ test/test_cuda_mixed_decimal_cmp.cu
@@ -1,0 +1,175 @@
+//  Copyright Matt Borland 2026.
+//  Use, modification and distribution are subject to the
+//  Boost Software License, Version 1.0. (See accompanying file
+//  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <iostream>
+#include <vector>
+#include <random>
+#include <boost/decimal/decimal32_t.hpp>
+#include <boost/decimal/decimal64_t.hpp>
+#include <boost/decimal/decimal128_t.hpp>
+#include "cuda_managed_ptr.hpp"
+#include "stopwatch.hpp"
+
+#include <cuda_runtime.h>
+
+using dec32_t = boost::decimal::decimal32_t;
+using dec64_t = boost::decimal::decimal64_t;
+using dec128_t = boost::decimal::decimal128_t;
+
+struct cmp_result
+{
+    bool eq;
+    bool neq;
+    bool lt;
+    bool le;
+    bool gt;
+    bool ge;
+};
+
+__global__ void cuda_test(const dec32_t *in32_lhs, const dec32_t *in32_rhs,
+                          const dec64_t *in64_lhs, const dec64_t *in64_rhs,
+                          const dec128_t *in128_lhs, const dec128_t *in128_rhs,
+                          cmp_result *out_32_64, cmp_result *out_32_128, cmp_result *out_64_128,
+                          int numElements)
+{
+    int i = blockDim.x * blockIdx.x + threadIdx.x;
+
+    if (i < numElements)
+    {
+        // dec32 vs dec64
+        out_32_64[i].eq  = (in32_lhs[i] == in64_lhs[i]);
+        out_32_64[i].neq = (in32_lhs[i] != in64_lhs[i]);
+        out_32_64[i].lt  = (in32_lhs[i] <  in64_lhs[i]);
+        out_32_64[i].le  = (in32_lhs[i] <= in64_lhs[i]);
+        out_32_64[i].gt  = (in32_lhs[i] >  in64_lhs[i]);
+        out_32_64[i].ge  = (in32_lhs[i] >= in64_lhs[i]);
+
+        // dec32 vs dec128
+        out_32_128[i].eq  = (in32_rhs[i] == in128_lhs[i]);
+        out_32_128[i].neq = (in32_rhs[i] != in128_lhs[i]);
+        out_32_128[i].lt  = (in32_rhs[i] <  in128_lhs[i]);
+        out_32_128[i].le  = (in32_rhs[i] <= in128_lhs[i]);
+        out_32_128[i].gt  = (in32_rhs[i] >  in128_lhs[i]);
+        out_32_128[i].ge  = (in32_rhs[i] >= in128_lhs[i]);
+
+        // dec64 vs dec128
+        out_64_128[i].eq  = (in64_rhs[i] == in128_rhs[i]);
+        out_64_128[i].neq = (in64_rhs[i] != in128_rhs[i]);
+        out_64_128[i].lt  = (in64_rhs[i] <  in128_rhs[i]);
+        out_64_128[i].le  = (in64_rhs[i] <= in128_rhs[i]);
+        out_64_128[i].gt  = (in64_rhs[i] >  in128_rhs[i]);
+        out_64_128[i].ge  = (in64_rhs[i] >= in128_rhs[i]);
+    }
+}
+
+int main(void)
+{
+    std::mt19937_64 rng{42};
+
+    int numElements = 50000;
+    std::cout << "[Mixed decimal comparison on " << numElements << " elements]" << std::endl;
+
+    cuda_managed_ptr<dec32_t> in32_lhs(numElements);
[…]
