### 2026-04-09 da0fb85
Fix unused variable warning
