### 2026-04-09 8120ce8
Fix -fpermissive warning on older GCC
