### 2026-04-21 d079af1
Add initial reproducer
