### 2026-04-21 65924d3
Fix integer rounding in ceil
