+++ include/boost/decimal/detail/cmath/ceil.hpp
@@ -60,6 +60,13 @@ constexpr auto ceil BOOST_DECIMAL_PREVENT_MACRO_SUBSTITUTION (const T val) noexc
     auto decimal_digits {static_cast<unsigned>(sig_dig)};
     const auto zero_digits {detail::remove_trailing_zeros(new_sig).number_of_removed_zeros};
     const auto non_zero_decimal_digits {decimal_digits - zero_digits};
+
+    if (non_zero_decimal_digits == 1 && (sig_dig + exp_ptr == 1))
+    {
+        // If the value is an integer, nothing should occur
+        return val;
+    }
+
     const auto non_zero_exp {exp_ptr + static_cast<int>(zero_digits)};
 
     if (non_zero_exp > static_cast<int>(non_zero_decimal_digits))
