### 2026-04-21 3d1ee48
Improve integer detection methods
