+++ include/boost/decimal/detail/cmath/floor.hpp
@@ -52,35 +52,25 @@ constexpr auto floor BOOST_DECIMAL_PREVENT_MACRO_SUBSTITUTION (const T val) noex
     const auto sig_dig {detail::precision_v<T>};
     auto decimal_digits {static_cast<unsigned>(sig_dig)};
     const auto zero_digits {detail::remove_trailing_zeros(new_sig).number_of_removed_zeros};
-    const auto non_zero_decimal_digits {decimal_digits - zero_digits};
+    const auto non_zero_exp {exp_ptr + static_cast<int>(zero_digits)};
 
-    if (non_zero_decimal_digits == 1 && (sig_dig + exp_ptr == 1))
+    if (non_zero_exp >= 0)
     {
         // If the value is an integer, nothing should occur
         return val;
     }
 
-    bool round {false};
-
     if (sig_dig > abs_exp)
     {
         decimal_digits = abs_exp;
-        if (sig_dig == abs_exp + 1)
-        {
-            round = true;
-        }
     }
     else if (exp_ptr < 1 && abs_exp >= sig_dig)
     {
         return is_neg ? neg_one : zero;
     }
-    else
-    {
-        --decimal_digits;
-    }
 
     new_sig /= detail::pow10<DivType>(decimal_digits);
-    if (is_neg && round)
+    if (is_neg)
     {
         ++new_sig;
     }
