### 2026-05-01 5c4d0cb
Remove unused vendored file given appearance of dependencies
