+++ include/boost/decimal/detail/int128/charconv.hpp
@@ -1,245 +1,0 @@
-// Copyright 2025 Matt Borland
-// Distributed under the Boost Software License, Version 1.0.
-// https://www.boost.org/LICENSE_1_0.txt
-//
-// This header injects additional functionality into the Boost.Charconv namespace
-// If this library is accepted into boost this functionality could be moved their
-
-#ifndef BOOST_DECIMAL_DETAIL_INT128_CHARCONV_HPP
-#define BOOST_DECIMAL_DETAIL_INT128_CHARCONV_HPP
-
-#if __has_include(<boost/charconv.hpp>)
-
-// Define for the user automatically,
-// otherwise we'll have an ever-increasing number of these required as we go down the dependency chain
-#if defined(BOOST_DECIMAL_DETAIL_INT128_ENABLE_CUDA) && !defined(BOOST_CHARCONV_ENABLE_CUDA)
-#  define BOOST_CHARCONV_ENABLE_CUDA
-#endif
-
-#include <boost/decimal/detail/int128/int128.hpp>
-#include <boost/decimal/detail/int128/literals.hpp>
-#include <boost/charconv.hpp>
-#include <boost/core/detail/string_view.hpp>
-
-namespace boost {
-namespace charconv {
-
-namespace detail {
-
-template <>
-struct is_signed<int128::uint128_t> { static constexpr bool value = false; };
-
-template <>
-struct is_signed<int128::int128_t> { static constexpr bool value = true; };
-
-template <>
-struct make_unsigned<int128::uint128_t> { using type = int128::uint128_t; };
-
-template <>
-struct make_unsigned<int128::int128_t> { using type = int128::uint128_t; };
-
-template <>
-struct make_signed<int128::uint128_t> { using type = int128::int128_t; };
-
-template <>
-struct make_signed<int128::int128_t> { using type = int128::int128_t; };
-
-#if defined(__CUDACC__) && defined(BOOST_DECIMAL_DETAIL_INT128_ENABLE_CUDA)
-
-template <>
-__host__ __device__ constexpr int128::uint128_t get_max_value<int128::uint128_t>()
-{
-    return std::numeric_limits<int128::uint128_t>::max();
-}
-
-template <>
-__host__ __device__ constexpr int128::int128_t get_max_value<int128::int128_t>()
-{
-    return std::numeric_limits<int128::int128_t>::max();
-}
-
-#endif // __NVCC__
-
-#if !(defined(__CUDACC__) && defined(BOOST_DECIMAL_DETAIL_INT128_ENABLE_CUDA))
-
-BOOST_DECIMAL_DETAIL_INT128_INLINE_CONSTEXPR int128::uint128_t int128_pow10[39] =
-{
-    int128::uint128_t{UINT64_C(0x0), UINT64_C(0x1)},
-    int128::uint128_t{UINT64_C(0x0), UINT64_C(0xa)},
-    int128::uint128_t{UINT64_C(0x0), UINT64_C(0x64)},
-    int128::uint128_t{UINT64_C(0x0), UINT64_C(0x3e8)},
-    int128::uint128_t{UINT64_C(0x0), UINT64_C(0x2710)},
-    int128::uint128_t{UINT64_C(0x0), UINT64_C(0x186a0)},
-    int128::uint128_t{UINT64_C(0x0), UINT64_C(0xf4240)},
-    int128::uint128_t{UINT64_C(0x0), UINT64_C(0x989680)},
[…]
