### 2026-05-02 f994f3a
More robust treatment and test of 1384
