### 2026-05-04 e708cab
Add minimal reproducer
