### 2026-05-04 06d413a
Add separate implementation
