+++ include/boost/decimal/detail/quantize_impl.hpp
@@ -1,0 +1,70 @@
+// Copyright 2026 Matt Borland
+// Distributed under the Boost Software License, Version 1.0.
+// https://www.boost.org/LICENSE_1_0.txt
+
+#ifndef BOOST_DECIMAL_DETAIL_QUANTIZE_IMPL_HPP
+#define BOOST_DECIMAL_DETAIL_QUANTIZE_IMPL_HPP
+
+#include <boost/decimal/fwd.hpp>
+#include <boost/decimal/detail/attributes.hpp>
+#include <boost/decimal/detail/concepts.hpp>
+#include <boost/decimal/detail/config.hpp>
+#include <boost/decimal/detail/fenv_rounding.hpp>
+#include <boost/decimal/detail/integer_search_trees.hpp>
+#include <boost/decimal/detail/power_tables.hpp>
+
+namespace boost {
+namespace decimal {
+namespace detail {
+
+// Implements the IEEE 754-2008 5.3.2 quantize operation by rescaling
+// the significand of x to match the quantum exponent of y.
+template <BOOST_DECIMAL_DECIMAL_FLOATING_TYPE DecimalType, typename Significand>
+BOOST_DECIMAL_CUDA_CONSTEXPR auto quantize_rescale(Significand& sig, const int delta, const bool sign) noexcept -> bool
+{
+    if (delta == 0 || sig == Significand{0})
+    {
+        return true;
+    }
+
+    if (delta > 0)
+    {
+        const auto sig_digits {num_digits(sig)};
+        if (sig_digits + delta > precision_v<DecimalType>)
+        {
+            return false;
+        }
+        sig *= pow10(static_cast<Significand>(delta));
+        return true;
+    }
+
+    const auto neg_delta {-delta};
+    const auto sig_digits {num_digits(sig)};
+
+    if (neg_delta > sig_digits)
+    {
+        const bool sticky {sig != Significand{0}};
+        sig = Significand{0};
+        fenv_round<DecimalType>(sig, sign, sticky);
+    }
+    else if (neg_delta > 1)
+    {
+        const auto pre_div {pow10(static_cast<Significand>(neg_delta - 1))};
+        const auto remainder {sig % pre_div};
+        sig /= pre_div;
+        const bool sticky {remainder != Significand{0}};
+        fenv_round<DecimalType>(sig, sign, sticky);
+    }
+    else
+    {
+        fenv_round<DecimalType>(sig, sign, false);
+    }
+
+    return true;
+}
+
+} // namespace detail
+} // namespace decimal
+} // namespace boost
+
+#endif // BOOST_DECIMAL_DETAIL_QUANTIZE_IMPL_HPP
