### 2026-05-04 24b986d
Add missing overloads
