+++ include/boost/decimal/cmath.hpp
@@ -201,6 +201,11 @@ BOOST_DECIMAL_EXPORT constexpr auto samequantum(decimal64_t lhs, decimal64_t rhs
     return samequantumd64(lhs, rhs);
 }
 
+BOOST_DECIMAL_EXPORT constexpr auto samequantum(decimal_fast64_t lhs, decimal_fast64_t rhs) noexcept -> bool
+{
+    return samequantumd64f(lhs, rhs);
+}
+
 BOOST_DECIMAL_EXPORT constexpr auto samequantum(decimal128_t lhs, decimal128_t rhs) noexcept -> bool
 {
     return samequantumd128(lhs, rhs);
@@ -254,6 +259,11 @@ BOOST_DECIMAL_EXPORT constexpr auto quantize(decimal_fast32_t lhs, decimal_fast3
 BOOST_DECIMAL_EXPORT constexpr auto quantize(decimal64_t lhs, decimal64_t rhs) noexcept -> decimal64_t
 {
     return quantized64(lhs, rhs);
+}
+
+BOOST_DECIMAL_EXPORT constexpr auto quantize(decimal_fast64_t lhs, decimal_fast64_t rhs) noexcept -> decimal_fast64_t
+{
+    return quantized64f(lhs, rhs);
 }
 
 BOOST_DECIMAL_EXPORT constexpr auto quantize(decimal128_t lhs, decimal128_t rhs) noexcept -> decimal128_t
