+++ include/boost/decimal/decimal_fast64_t.hpp
@@ -26,6 +26,7 @@
 #include <boost/decimal/detail/to_decimal.hpp>
 #include <boost/decimal/detail/promotion.hpp>
 #include <boost/decimal/detail/check_non_finite.hpp>
+#include <boost/decimal/detail/quantize_impl.hpp>
 #include <boost/decimal/detail/shrink_significand.hpp>
 #include <boost/decimal/detail/cmath/isfinite.hpp>
 #include <boost/decimal/detail/cmath/fpclassify.hpp>
@@ -500,7 +501,9 @@ public:
     friend constexpr auto copysignd64f(decimal_fast64_t mag, decimal_fast64_t sgn) noexcept -> decimal_fast64_t;
     friend constexpr auto scalbnd64f(decimal_fast64_t num, int exp) noexcept -> decimal_fast64_t;
     friend constexpr auto scalblnd64f(decimal_fast64_t num, long exp) noexcept -> decimal_fast64_t;
+    friend constexpr auto samequantumd64f(decimal_fast64_t lhs, decimal_fast64_t rhs) noexcept -> bool;
     friend constexpr auto quantexpd64f(decimal_fast64_t x) noexcept -> int;
+    friend constexpr auto quantized64f(decimal_fast64_t lhs, decimal_fast64_t rhs) noexcept -> decimal_fast64_t;
 };
 
 #ifdef _MSC_VER
@@ -1662,6 +1665,29 @@ constexpr auto decimal_fast64_t::operator--(int) noexcept -> decimal_fast64_t
     return temp;
 }
 
+// Effects: determines if the quantum exponents of x and y are the same.
+// If both x and y are NaN, or infinity, they have the same quantum exponents;
+// if exactly one operand is infinity or exactly one operand is NaN, they do not have the same quantum exponents.
+// The samequantum functions raise no exception.
+constexpr auto samequantumd64f(const decimal_fast64_t lhs, const decimal_fast64_t rhs) noexcept -> bool
+{
+    #ifndef BOOST_DECIMAL_FAST_MATH
+    const auto lhs_fp {fpclassify(lhs)};
+    const auto rhs_fp {fpclassify(rhs)};
+
+    if ((lhs_fp == FP_NAN && rhs_fp == FP_NAN) || (lhs_fp == FP_INFINITE && rhs_fp == FP_INFINITE))
+    {
+        return true;
+    }
+    if ((lhs_fp == FP_NAN || rhs_fp == FP_INFINITE) || (rhs_fp == FP_NAN || lhs_fp == FP_INFINITE))
+    {
+        return false;
+    }
+    #endif
+
+    return lhs.unbiased_exponent() == rhs.unbiased_exponent();
+}
+
 // Effects: if x is finite, returns its quantum exponent.
 // Otherwise, a domain error occurs and INT_MIN is returned.
 constexpr auto quantexpd64f(const decimal_fast64_t x) noexcept -> int
@@ -1674,6 +1700,55 @@ constexpr auto quantexpd64f(const decimal_fast64_t x) noexcept -> int
     #endif
 
     return static_cast<int>(x.unbiased_exponent());
+}
+
+// Returns: a number that is equal in value (except for any rounding) and sign to x,
+// and which has an exponent set to be equal to the exponent of y.
+// If the exponent is being increased, the value is correctly rounded according to the current rounding mode;
+// if the result does not have the same value as x, the "inexact" floating-point exception is raised.
+// If the exponent is being decreased and the significand of the result has more digits than the type would allow,
+// the "invalid" floating-point exception is raised and the result is NaN.
+// If one or both operands are NaN, the result is NaN.
[…]
