### 2026-05-04 f4f93e6
Consolidate now that we have shared impl
