+++ include/boost/decimal/decimal_fast32_t.hpp
@@ -498,7 +498,8 @@ public:
     // Specific decimal functionality
     friend constexpr auto samequantumd32f(decimal_fast32_t lhs, decimal_fast32_t rhs) noexcept -> bool;
     friend constexpr auto quantexpd32f(decimal_fast32_t x) noexcept -> int;
-    friend constexpr auto quantized32f(decimal_fast32_t lhs, decimal_fast32_t rhs) noexcept -> decimal_fast32_t;
+    template <BOOST_DECIMAL_DECIMAL_FLOATING_TYPE T>
+    friend BOOST_DECIMAL_CUDA_CONSTEXPR auto quantize(T lhs, T rhs) noexcept -> T;
 };
 
 #ifdef _MSC_VER
@@ -1683,55 +1684,6 @@ constexpr auto quantexpd32f(const decimal_fast32_t x) noexcept -> int
     #endif
 
     return static_cast<int>(x.unbiased_exponent());
-}
-
-// Returns: a number that is equal in value (except for any rounding) and sign to x,
-// and which has an exponent set to be equal to the exponent of y.
-// If the exponent is being increased, the value is correctly rounded according to the current rounding mode;
-// if the result does not have the same value, as x, the "inexact" floating-point exception is raised.
-// If the exponent is being decreased and the significand of the result has more digits than the type would allow,
-// the "invalid" floating-point exception is raised and the result is NaN.
-// If one or both operands are NaN, the result is NaN.
-// Otherwise, if only one operand is infinity, the "invalid" floating-point exception is raised and the result is NaN.
-// If both operands are infinity, the result is DEC_INFINITY, with the same sign as x, converted to the type of x.
-// The quantize functions do not signal underflow.
-constexpr auto quantized32f(const decimal_fast32_t lhs, const decimal_fast32_t rhs) noexcept -> decimal_fast32_t
-{
-    #ifndef BOOST_DECIMAL_FAST_MATH
-    // Return the correct type of nan
-    if (isnan(lhs))
-    {
-        return lhs;
-    }
-    else if (isnan(rhs))
-    {
-        return rhs;
-    }
-
-    // If one is infinity then return a signaling NAN
-    if (isinf(lhs) != isinf(rhs))
-    {
-        return direct_init(detail::d32_fast_snan, UINT8_C(0));
-    }
-    else if (isinf(lhs) && isinf(rhs))
-    {
-        return lhs;
-    }
-    #endif
-
-    auto components {lhs.to_components()};
-    const auto rhs_exp {rhs.biased_exponent()};
-
-    if (!detail::quantize_rescale<decimal_fast32_t>(components.sig, components.exp - rhs_exp, components.sign))
-    {
-        #ifndef BOOST_DECIMAL_FAST_MATH
-        return direct_init(detail::d32_fast_snan, UINT8_C(0));
-        #else
-        return {components.sig, rhs_exp, components.sign};
-        #endif
-    }
-
-    return {components.sig, rhs_exp, components.sign};
 }
 
 #if !defined(BOOST_DECIMAL_DISABLE_CLIB)
