+++ include/boost/decimal/detail/quantize_impl.hpp
@@ -12,6 +12,12 @@
 #include <boost/decimal/detail/fenv_rounding.hpp>
 #include <boost/decimal/detail/integer_search_trees.hpp>
 #include <boost/decimal/detail/power_tables.hpp>
+#include <boost/decimal/detail/cmath/isfinite.hpp>
+#include <boost/decimal/detail/cmath/fpclassify.hpp>
+
+#ifndef BOOST_DECIMAL_BUILD_MODULE
+#include <limits>
+#endif
 
 namespace boost {
 namespace decimal {
@@ -64,6 +70,58 @@ BOOST_DECIMAL_CUDA_CONSTEXPR auto quantize_rescale(Significand& sig, const int d
 }
 
 } // namespace detail
+
+// IEEE 754-2008 3.6.6
+// Returns: a number that is equal in value (except for any rounding) and sign to lhs,
+// and which has an exponent set to be equal to the exponent of rhs.
+// If the exponent is being increased, the value is correctly rounded according to the current rounding mode;
+// if the result does not have the same value as lhs, the "inexact" floating-point exception is raised.
+// If the exponent is being decreased and the significand of the result has more digits than the type would allow,
+// the "invalid" floating-point exception is raised and the result is NaN.
+// If one or both operands are NaN, the result is NaN.
+// Otherwise, if only one operand is infinity, the "invalid" floating-point exception is raised and the result is NaN.
+// If both operands are infinity, the result is DEC_INFINITY, with the same sign as lhs.
+// The quantize functions do not signal underflow.
+BOOST_DECIMAL_EXPORT template <BOOST_DECIMAL_DECIMAL_FLOATING_TYPE DecimalType>
+BOOST_DECIMAL_CUDA_CONSTEXPR auto quantize(const DecimalType lhs, const DecimalType rhs) noexcept -> DecimalType
+{
+    #ifndef BOOST_DECIMAL_FAST_MATH
+    // Return the correct type of nan
+    if (isnan(lhs))
+    {
+        return lhs;
+    }
+    if (isnan(rhs))
+    {
+        return rhs;
+    }
+
+    // If exactly one operand is infinity then return a signaling NaN
+    if (isinf(lhs) != isinf(rhs))
+    {
+        return std::numeric_limits<DecimalType>::signaling_NaN();
+    }
+    if (isinf(lhs) && isinf(rhs))
+    {
+        return lhs;
+    }
+    #endif
+
+    auto components {lhs.to_components()};
+    const auto rhs_exp {rhs.biased_exponent()};
+
+    if (!detail::quantize_rescale<DecimalType>(components.sig, components.exp - rhs_exp, components.sign))
+    {
+        #ifndef BOOST_DECIMAL_FAST_MATH
+        return std::numeric_limits<DecimalType>::signaling_NaN();
+        #else
+        return {components.sig, rhs_exp, components.sign};
+        #endif
+    }
+
+    return {components.sig, rhs_exp, components.sign};
+}
+
 } // namespace decimal
 } // namespace boost
 
