### 2026-05-04 1fff1c6
Fix nan handling
