### 2026-05-04 778683d
SNAN must payload must be returned
