### 2026-05-06 e026f05
Fix constructor handling of 0
