+++ include/boost/decimal/decimal32_t.hpp
@@ -642,7 +642,18 @@ BOOST_DECIMAL_CUDA_CONSTEXPR decimal32_t::decimal32_t(T1 coeff, T2 exp, const de
 
     if (reduced_coeff == 0U)
     {
-        // Normalize our handling of zeros
+        // IEEE 754-2008 3.5.1: zero has a cohort with one representation per exponent.
+        // Clamp the requested exponent to the representable range and encode it; sign was already set.
+        auto zero_biased_exp {biased_exp};
+        if (zero_biased_exp < 0)
+        {
+            zero_biased_exp = 0;
+        }
+        else if (zero_biased_exp > static_cast<int>(detail::d32_max_biased_exponent))
+        {
+            zero_biased_exp = static_cast<int>(detail::d32_max_biased_exponent);
+        }
+        bits_ |= (static_cast<std::uint32_t>(zero_biased_exp) << detail::d32_not_11_exp_shift) & detail::d32_not_11_exp_mask;
         return;
     }
 
