+++ include/boost/decimal/decimal_fast128_t.hpp
@@ -522,13 +522,32 @@ constexpr decimal_fast128_t::decimal_fast128_t(T1 coeff, T2 exp, const detail::c
     const auto is_negative {static_cast<bool>(resultant_sign)};
     sign_ = is_negative;
 
+    // IEEE 754-2008 3.5.1: zero has a cohort with one representation per exponent.
+    // Skip normalization for zero (which would otherwise expand the significand and shift the exponent)
+    // and clamp the requested exponent to the representable range.
+    if (min_coeff == minimum_coefficient_size{0})
+    {
+        significand_ = static_cast<significand_type>(0);
+        auto biased_exp {static_cast<int>(exp) + detail::bias_v<decimal_fast128_t>};
+        if (biased_exp < 0)
+        {
+            biased_exp = 0;
+        }
+        else if (biased_exp > detail::max_biased_exp_v<decimal_fast128_t>)
+        {
+            biased_exp = detail::max_biased_exp_v<decimal_fast128_t>;
+        }
+        exponent_ = static_cast<exponent_type>(biased_exp);
+        return;
+    }
+
     // Normalize the significand in the constructor, so we don't have
     // to calculate the number of digits for operations
     detail::normalize<decimal_fast128_t>(min_coeff, exp, is_negative);
 
     significand_ = static_cast<significand_type>(min_coeff);
 
-    const auto biased_exp {significand_ == 0U ? 0 : exp + detail::bias_v<decimal_fast128_t>};
+    const auto biased_exp {static_cast<int>(exp) + detail::bias_v<decimal_fast128_t>};
 
     if (biased_exp > detail::max_biased_exp_v<decimal_fast128_t>)
     {
@@ -540,10 +559,9 @@ constexpr decimal_fast128_t::decimal_fast128_t(T1 coeff, T2 exp, const detail::c
     }
     else
     {
-        // Flush denorms to zero
+        // Flush denorms to zero, preserving sign per IEEE 754-2008 3.5.1
         significand_ = static_cast<significand_type>(0);
-        exponent_ = static_cast<exponent_type>(detail::bias_v<decimal_fast128_t>);
-        sign_ = false;
+        exponent_ = static_cast<exponent_type>(0);
     }
 }
 
