### 2026-05-06 5d06258
Special case 0 in charconv handling
