### 2026-05-06 e2636a5
Remove normalized sign handling
