### 2026-05-06 419a71c
Update tests for new handling of zero
