+++ acm/paper/toms_decimal.tex
@@ -1,0 +1,1154 @@
+%%
+%% Boost.Decimal: outline of a research paper for ACM Transactions on
+%% Mathematical Software (TOMS).
+%%
+%% This is a complete first draft: every section is written in prose. The
+%% remaining camera-ready placeholders are the rights/DOI/volume block, the CCS
+%% concept codes, the received dates, and final verification of references.bib.
+%%
+%% It uses the unmodified ACM template in ../template (acmart.cls and
+%% ACM-Reference-Format.bst). Build with the provided Makefile, which points
+%% LaTeX at that folder without changing it. See README.md.
+%%
+%% TOMS is published in the "acmsmall" journal format.
+%%
+\documentclass[acmsmall]{acmart}
+
+\AtBeginDocument{%
+  \providecommand\BibTeX{{%
+    Bib\TeX}}}
+
+%% Code listings: compact, monochrome C++ style. Standard listings package.
+\usepackage{listings}
+\lstset{
+  language=C++,
+  basicstyle=\footnotesize\ttfamily,
+  columns=fullflexible,
+  keepspaces=true,
+  showstringspaces=false,
+  breaklines=true,
+  frame=single,
+  framesep=4pt,
+  xleftmargin=4pt,
+  belowcaptionskip=4pt,
+  aboveskip=6pt,
+  belowskip=4pt
+}
+
+%% Rights: SAMPLE values. Replace with the values from the ACM rights form
+%% once the paper is accepted.
+\setcopyright{acmlicensed}
+\copyrightyear{2026}
+\acmYear{2026}
+\acmDOI{XXXXXXX.XXXXXXX}
+
+%% Journal: TOMS.
+\acmJournal{TOMS}
+\acmVolume{0}
+\acmNumber{0}
+\acmArticle{0}
+\acmMonth{0}
+
+%% Title. Keep a short title for the running header.
+\title[Boost.Decimal: Portable IEEE 754 Decimal Floating-Point]{Boost.Decimal:
+  A Portable, Header-Only, constexpr and GPU-Capable Implementation of
+  IEEE 754 Decimal Floating-Point Arithmetic}
+
+%% Authors. ORCID for Matt Borland is taken from the JOSS submission.
+\author{Matt Borland}
+\orcid{0009-0005-8183-4254}
+\affiliation{%
+  \institution{The C++ Alliance}
+  \country{USA}}
+\email{matt@mattborland.com}
+
+\author{Christopher Kormanyos}
+\affiliation{%
+  \institution{Independent Researcher}
+  \country{Germany}}
+\email{e\_float@yahoo.com}
+
+\renewcommand{\shortauthors}{Borland and Kormanyos}
+
+%%
+%% Abstract. Draft below; tighten once results are final.
+\begin{abstract}
+  Decimal floating-point arithmetic, standardized in IEEE 754, represents
+  values such as 0.1 exactly and is required wherever results must match
+  base-ten expectations, most notably in finance and commerce. The two
+  established implementations, the Intel Decimal Floating-Point Math Library
+  and IBM decNumber, are C libraries: they are not header-only, expose no
+  C++ compile-time interface, do not run on accelerators, and do not cover
+  the full range of CPU architectures used today. We present Boost.Decimal,
+  a header-only C++14 implementation of the IEEE 754 decimal32, decimal64,
+  and decimal128 types (and three non-conforming variants tuned for runtime
+  speed) that requires no external dependency, evaluates at compile time
+  through constexpr, executes on CPUs and on NVIDIA GPUs from a single source,
+  and is validated for cross-platform bit-for-bit consistency across x86,
+  ARM, s390x, PPC64LE, and bare-metal targets. We describe the design, the
+  arithmetic algorithms, the conformance and reproducibility testing, and a
[…]
