+++ acm/paper/references.bib
@@ -1,0 +1,243 @@
+%% Seed bibliography for the Boost.Decimal TOMS paper outline.
+%% Entries are real and verifiable; complete page numbers, DOIs, and access
+%% dates before the camera-ready. Several entries are reused from the JOSS
+%% submission's paper.bib.
+
+@article{IEEE754-2008,
+  key     = {IEEE},
+  author  = {{IEEE}},
+  title   = {IEEE Standard for Floating-Point Arithmetic},
+  journal = {IEEE Std 754-2008},
+  year    = {2008},
+  pages   = {1-70},
+  doi     = {10.1109/IEEESTD.2008.4610935}
+}
+
+@article{IEEE754-2019,
+  key     = {IEEE},
+  author  = {{IEEE}},
+  title   = {IEEE Standard for Floating-Point Arithmetic},
+  journal = {IEEE Std 754-2019 (Revision of IEEE 754-2008)},
+  year    = {2019},
+  pages   = {1-84},
+  doi     = {10.1109/IEEESTD.2019.8766229}
+}
+
+@inproceedings{Cowlishaw2003,
+  author    = {Cowlishaw, Michael F.},
+  title     = {Decimal Floating-Point: Algorism for Computers},
+  booktitle = {Proceedings of the 16th IEEE Symposium on Computer Arithmetic (ARITH-16)},
+  year      = {2003},
+  doi       = {10.1109/ARITH.2003.1207666},
+  url       = {https://speleotrove.com/decimal/IEEE-cowlishaw-arith16.pdf}
+}
+
+@article{cornea2009,
+  author  = {Cornea, Marius and Harrison, John and Anderson, Cristina and Tang, Ping Tak Peter and Schneider, Eric and Gvozdev, Evgeny},
+  title   = {A Software Implementation of the {IEEE} 754R Decimal Floating-Point Arithmetic Using the Binary Encoding Format},
+  journal = {IEEE Transactions on Computers},
+  volume  = {58},
+  number  = {2},
+  pages   = {148-162},
+  year    = {2009},
+  doi     = {10.1109/TC.2008.209},
+  url     = {https://www.cl.cam.ac.uk/~jrh13/papers/decimal.pdf}
+}
+
+@misc{intel-dfp,
+  author       = {{Intel Corporation}},
+  title        = {Intel Decimal Floating-Point Math Library},
+  howpublished = {\url{https://www.intel.com/content/www/us/en/developer/articles/tool/intel-decimal-floating-point-math-library.html}},
+  year         = {2023},
+  note         = {Accessed 2026. Software implementation of IEEE 754-2008 decimal arithmetic}
+}
+
+@misc{ibm-decnumber,
+  author       = {Cowlishaw, Michael F. and {IBM Corporation}},
+  title        = {The {decNumber} Library and General Decimal Arithmetic},
+  howpublished = {\url{https://speleotrove.com/decimal/}},
+  year         = {2010},
+  note         = {Portable ANSI C reference implementation; used by GCC and ICU}
+}
+
+@misc{tr24733,
+  author       = {{ISO/IEC}},
+  title        = {{ISO/IEC} {TR} 24733: Extension for the Programming Language {C++} to Support Decimal Floating-Point Arithmetic},
+  howpublished = {\url{https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2009/n2849.pdf}},
+  year         = {2011}
+}
+
+@misc{wg21-n3407,
+  author       = {Klarer, Robert and others},
+  title        = {{N3407}: Proposal for Adding Decimal Floating Point Support to {C++}},
+  howpublished = {\url{https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2012/n3407.html}},
+  year         = {2012},
[…]
