### 2026-06-30 c4b6c65
Add B2 debug configuration flag
