+++ .github/workflows/qemu.yml
@@ -189,6 +189,7 @@ jobs:
           B2_CXXSTD: ${{matrix.cxxstd}}
           B2_SANITIZE: ${{matrix.sanitize}}
           B2_STDLIB: ${{matrix.stdlib}}
+          B2_FLAGS: --debug-configuration
           # More entries can be added in the same way, see the B2_ARGS assignment in ci/enforce.sh for the possible keys.
           B2_DEFINES: 'BOOST_DECIMAL_QEMU_TEST'
           # Variables set here (to non-empty) will override the top-level environment variables, e.g.
