+++ paper/paper.bib
@@ -40,8 +40,25 @@
 @inproceedings{Cowlishaw:2003,
   author = {Cowlishaw, Michael F.},
   title = {Decimal Floating-Point: Algorism for Computers},
+  booktitle = {Proceedings of the 16th IEEE Symposium on Computer Arithmetic (ARITH-16)},
+  series = {ARITH-16},
+  address = {Santiago de Compostela, Spain},
+  publisher = {IEEE Computer Society},
+  pages = {104--111},
   month = {June},
   year = {2003},
   doi = {10.1109/ARITH.2003.1207666},
   url = {https://speleotrove.com/decimal/IEEE-cowlishaw-arith16.pdf}
+}
+
+@article{Goldberg:1991,
+  author = {Goldberg, David},
+  title = {What Every Computer Scientist Should Know About Floating-Point Arithmetic},
+  journal = {ACM Computing Surveys},
+  volume = {23},
+  number = {1},
+  pages = {5--48},
+  year = {1991},
+  publisher = {Association for Computing Machinery},
+  doi = {10.1145/103162.103163}
 }
