### 2026-08-10 8c4ac81
Make the code examples proper figures
