### 2026-08-11 06e8866
Change clang version number
Ref: https://github.com/openjournals/joss-reviews/issues/10345#issuecomment-5248321918
