### 2026-08-12 7cb704a
minor change in paper.md
