### 2026-08-12 bdf2fa3
Add Citation section to front page of docs
