### 2026-08-25 #10345 @arfon
"Fixed" in https://github.com/openjournals/joss/pull/1570
