### 2026-09-04 23c9b19
Make the fma of the 128 bit types round one time
fma must calculate x * y + z with one rounding, which is the definition of
fusedMultiplyAdd in IEEE 754-2019. The d128_fma_impl function calculated that
expression with the operators of the type, thus it rounded two times. Of 20000
random triples, 16.14% were wrong and the largest error was 118 units of the
last place. When z was -(x * y) the result was 0 for every pair, and the correct
value there is the rounding error of the product. Four functions took that
path: fma and unchecked_fma, of decimal128_t and of decimal_fast128_t.

The 32 and 64 
[…]
