+++ test/test_fma.cpp
@@ -1,0 +1,467 @@
+// Copyright 2026 Matt Borland
+// Distributed under the Boost Software License, Version 1.0.
+// https://www.boost.org/LICENSE_1_0.txt
+//
+// fma must calculate x * y + z with one rounding (IEEE 754-2019 fusedMultiplyAdd).
+// The expected values come from MPFR at 900 bits, at the 34 digits of decimal128_t.
+
+#include <boost/decimal.hpp>
+#include <boost/core/lightweight_test.hpp>
+#include <cstring>
+#include <iostream>
+#include <limits>
+#include <random>
+
+using namespace boost::decimal;
+
+struct fma_case
+{
+    const char* x;
+    const char* y;
+    const char* z;
+    const char* expected;
+};
+
+static const fma_case fma_cases[] =
+{
+// the three worst triples of the report
+    { "4.891128916954585116359252097589272e+03",
+      "-7.570510793333322422709743010426965e-02",
+      "3.707536042839128637930999007415480e+02",
+      "4.701617100187612469012746705158118e-01" },
+    { "9.773367626875150075486696879175602e-02",
+      "-3.337907954067831540381441089089419e-01",
+      "3.235029480287355880201961692407601e-02",
+      "-2.723067369020513807361798191291154e-04" },
+    { "-1.635908948849659232280107483909349e+02",
+      "-3.176446739966658464358952445855542e-03",
+      "-5.281416298488918666313435068678353e-01",
+      "-8.503865103313556434996443805914455e-03" },
+// random triples, exponents of 10^-3 to 10^3
+    { "-1.065008300110031905957897280440974e-03",
+      "8.679786286031742647482469037679398e-01",
+      "2.459772220945470025638742347671538e-01",
+      "2.450528176507664992227350252829746e-01" },
+    { "2.204698774688623268682332486262565e+02",
+      "4.391348326395192744337294837450587e-01",
+      "6.616668522588143528118897915024711e+03",
+      "6.713484525332487707332688687817646e+03" },
+    { "-9.976911973353208647336163725677450e-01",
+      "2.169266061245795773782627422211721e-03",
+      "-7.439068330870974128742758471652387e+00",
+      "-7.441232588524957322211015428677276e+00" },
+    { "-1.888740352881831634403692216579063e+01",
+      "-7.900118649737038306011084617881166e+02",
+      "-4.999754949433110540456096531467187e+01",
+      "1.487127533681834187300900128042427e+04" },
+    { "8.171530766625406270093349670323739e+03",
+      "2.593580490537856789189978389609014e-03",
+      "-5.576232872708608952707412919282934e-01",
+      "2.063589948687864924607130452970817e+01" },
+    { "-7.750345597256329137563280171352258e-02",
+      "-9.100287754281406319757676608288459e+02",
+      "3.073392767787947858334881593235580e+02",
+      "3.778696519089553701144226749142327e+02" },
+    { "6.013317852722187956978891138409258e-03",
+      "1.570765883557867700949175676275650e-01",
+      "-3.229835276084885365519028711089201e-02",
+      "-3.135380130784430692263921963834903e-02" },
+    { "9.033523195572541015573502066498907e+02",
+      "-2.046124559993565407347510599148056e+01",
+      "4.920017356632420627073578236505352e-03",
+      "-1.848370875371517575917133434459618e+04" },
+    { "7.517577000791313867554357371910665e+00",
+      "8.135491485969028894823460725489820e-03",
+      "7.227824998636771650829718582341538e+00",
[…]
