+++ include/boost/decimal/detail/cmath/fma.hpp
@@ -11,6 +11,10 @@
 #include <boost/decimal/decimal128_t.hpp>
 #include <boost/decimal/decimal_fast128_t.hpp>
 #include <boost/decimal/detail/config.hpp>
+#include <boost/decimal/detail/i256.hpp>
+#include <boost/decimal/detail/normalize.hpp>
+#include <boost/decimal/detail/power_tables.hpp>
+#include <boost/decimal/detail/u256.hpp>
 
 namespace boost {
 namespace decimal {
@@ -96,10 +100,202 @@ constexpr auto d64_fma_impl(T x, T y, T z) noexcept -> T
     return detail::d128_add_impl_new<T>(first_res, z_components);
 }
 
-template <bool, BOOST_DECIMAL_DECIMAL_FLOATING_TYPE T>
+// 10^38 is the largest power of ten which 128 bits holds. Every divisor of the alignment
+// stays at this width, thus u256 takes its narrow paths.
+BOOST_DECIMAL_INLINE_CONSTEXPR_VARIABLE std::int32_t fma_narrow_digits {38};
+
+// Marks a quotient which dropped a digit that is not zero: a last digit of zero becomes a
+// one. The mark stays far below the last digit of the result, thus the rounding is exact.
+BOOST_DECIMAL_CUDA_CONSTEXPR auto fma_apply_sticky(const u256& quotient, const bool inexact) noexcept -> u256
+{
+    if (inexact && quotient % UINT64_C(10) == u256{0, 0, 0, 0})
+    {
+        return quotient + u256{0, 0, 0, 1};
+    }
+
+    return quotient;
+}
+
+// Divides by a power of ten and reports a dropped digit which is not zero. A shift of more
+// than fma_narrow_digits goes in steps, because a narrow divisor costs less than a wide one.
+BOOST_DECIMAL_CUDA_CONSTEXPR auto fma_shift_right(const u256& sig, std::int32_t shift, const std::int32_t width, bool& inexact) noexcept -> u256
+{
+    if (shift > width)
+    {
+        // Every digit falls below the sticky position, and the value is not zero.
+        inexact = false;
+        return u256{0, 0, 0, 1};
+    }
+
+    u256 quotient {sig};
+    inexact = false;
+
+    while (shift > 0)
+    {
+        const auto step {shift < fma_narrow_digits ? shift : fma_narrow_digits};
+        const auto result {detail::impl::div_mod(quotient, detail::pow10_u128(static_cast<std::size_t>(step)))};
+
+        inexact = inexact || result.remainder != u256{0, 0, 0, 0};
+        quotient = result.quotient;
+        shift -= step;
+    }
+
+    return quotient;
+}
+
+// Narrows the sum to the 38 digits of a 128 bit significand, with the sticky mark in the
+// last place. The constructor then rounds those digits one time.
+BOOST_DECIMAL_CUDA_CONSTEXPR auto fma_narrow_sum(const u256& sig, std::int32_t& exp) noexcept -> int128::uint128_t
+{
+    const std::int32_t digits {detail::num_digits(sig)};
+
+    if (digits <= fma_narrow_digits)
+    {
+        return int128::uint128_t{sig[1], sig[0]};
+    }
+
+    const auto shift {digits - fma_narrow_digits};
+    bool inexact {};
+    const auto shifted {fma_shift_right(sig, shift, digits, inexact)};
+    exp += shift;
+
+    // The quotient holds 38 digits, thus the mark goes on in 128 bits and not in 256.
+    int128::uint128_t narrow {shifted[1], shifted[0]};
+
+    if (inexact && narrow % UINT64_C(10) == 0U)
+    {
+        ++narrow;
+    }
+
+    return narrow;
+}
+
[…]
