### 2026-09-06 33fd17b
Refactor inverse hyperbolic trig funcs
