### 2026-09-06 bc7566f
Review comments and compile warnings
