+++ include/boost/decimal/detail/cmath/asinh.hpp
@@ -45,37 +45,46 @@ constexpr auto asinh_impl(const T x) noexcept
         else if (x > zero)
         {
             constexpr T tenth_root_epsilon { exp(log(std::numeric_limits<T>::epsilon()) / 10) };
+            constexpr T four_tenths { 4, -1 };
 
             const auto xsq = x * x;
 
-            if (x > one / tenth_root_epsilon)
+            constexpr T asymp_x { one / tenth_root_epsilon };
+
+            if (x > asymp_x)
             {
                 // http://functions.wolfram.com/ElementaryFunctions/ArcSinh/06/01/06/01/0001/
                 // approximation by Laurent series in 1/x at 0+ order from -1 to 9
                 const auto inv_xsq = one / xsq;
+
+                constexpr T one_fourth { T { 1, 0 } / T { 4, 0 } };
+                constexpr T minus_three_over_32 { -T { 3, 0 } / T { 32, 0 } };
+                constexpr T five_over_96 { T { 5, 0 } / T { 96, 0 } };
+                constexpr T minus_thirty_five_over_1024 { -T { 35, 0 } / T { 1024, 0 } };
+                constexpr T sixty_three_over_2560 { T { 63, 0 } / T { 2560, 0 } };
 
                 result =
                     numbers::ln2_v<T>
                   + ::boost::decimal::log(x)
                   + inv_xsq *
                     (
-                        one / T { 4, 0 }
+                        one_fourth
                       + inv_xsq *
                         (
-                            -T { 3, 0 } / T { 32, 0 }
+                            minus_three_over_32
                           + inv_xsq *
                             (
-                                T { 5, 0 } / T { 96, 0 }
+                                five_over_96
                               + inv_xsq *
                                 (
-                                    -T { 35, 0 } / T { 1024, 0 }
-                                  + inv_xsq * (T { 63, 0 } / T { 2560, 0 })
+                                    minus_thirty_five_over_1024
+                                  + inv_xsq * sixty_three_over_2560
                                 )
                             )
                         )
                     );
             }
-            else if(x >= T { 4 , -1 })
+            else if(x >= four_tenths)
             {
                 // http://functions.wolfram.com/ElementaryFunctions/ArcSinh/02/
                 result = ::boost::decimal::log(x + sqrt(xsq + one));
@@ -93,7 +102,12 @@ constexpr auto asinh_impl(const T x) noexcept
                 // Normal[Series[ArcSinh[x], {x, 0, 9}]]
                 // FullSimplify[%]
                 // HornerForm[%]
-                result = x * (1 + xsq * (-(one/6) + xsq * (T { 3, 0 } / 40 + xsq * (-(T { 5, 0 } / 112) + (35 * xsq) / 1152))));
+                constexpr T minus_one_sixth { -T { 1, 0 } / T { 6, 0 } };
+                constexpr T three_over_40 { T { 3, 0 } / T { 40, 0 } };
+                constexpr T minus_five_over_112 { -T { 5, 0 } / T { 112, 0 } };
+                constexpr T thirty_five_over_1152 { T { 35, 0 } / T { 1152, 0 } };
+
+                result = x * (one + xsq * (minus_one_sixth + xsq * (three_over_40 + xsq * (minus_five_over_112 + thirty_five_over_1152 * xsq))));
             }
         }
     }
