### 2026-09-07 98d2689
Improve inverse hyper trig perf
