+++ test/test_acosh.cpp
@@ -392,12 +392,12 @@ auto main() -> int
   {
     using namespace boost::decimal;
 
-    const bool result_acosh_dec32_t_is_ok       = local::test_acosh_ctrl<decimal32_t>      (16);
-    const bool result_acosh_dec64_t_is_ok       = local::test_acosh_ctrl<decimal64_t>      (16);
-    const bool result_acosh_dec128_t_is_ok      = local::test_acosh_ctrl<decimal128_t>     (16);
-    const bool result_acosh_dec_fast32_t_is_ok  = local::test_acosh_ctrl<decimal_fast32_t> (16);
-    const bool result_acosh_dec_fast64_t_is_ok  = local::test_acosh_ctrl<decimal_fast64_t> (16);
-    const bool result_acosh_dec_fast128_t_is_ok = local::test_acosh_ctrl<decimal_fast128_t>(16);
+    const bool result_acosh_dec32_t_is_ok       = local::test_acosh_ctrl<decimal32_t>      (8);
+    const bool result_acosh_dec64_t_is_ok       = local::test_acosh_ctrl<decimal64_t>      (8);
+    const bool result_acosh_dec128_t_is_ok      = local::test_acosh_ctrl<decimal128_t>     (8);
+    const bool result_acosh_dec_fast32_t_is_ok  = local::test_acosh_ctrl<decimal_fast32_t> (8);
+    const bool result_acosh_dec_fast64_t_is_ok  = local::test_acosh_ctrl<decimal_fast64_t> (8);
+    const bool result_acosh_dec_fast128_t_is_ok = local::test_acosh_ctrl<decimal_fast128_t>(8);
 
     BOOST_TEST(result_acosh_dec32_t_is_ok);
     BOOST_TEST(result_acosh_dec64_t_is_ok);
@@ -424,12 +424,12 @@ auto main() -> int
   {
     using namespace boost::decimal;
 
-    const bool result_acosh_ranges_dec32_t_is_ok       = local::test_acosh_ranges<decimal32_t>      (16);
-    const bool result_acosh_ranges_dec64_t_is_ok       = local::test_acosh_ranges<decimal64_t>      (16);
-    const bool result_acosh_ranges_dec128_t_is_ok      = local::test_acosh_ranges<decimal128_t>     (16);
-    const bool result_acosh_ranges_dec_fast32_t_is_ok  = local::test_acosh_ranges<decimal_fast32_t> (16);
-    const bool result_acosh_ranges_dec_fast64_t_is_ok  = local::test_acosh_ranges<decimal_fast64_t> (16);
-    const bool result_acosh_ranges_dec_fast128_t_is_ok = local::test_acosh_ranges<decimal_fast128_t>(16);
+    const bool result_acosh_ranges_dec32_t_is_ok       = local::test_acosh_ranges<decimal32_t>      (8);
+    const bool result_acosh_ranges_dec64_t_is_ok       = local::test_acosh_ranges<decimal64_t>      (8);
+    const bool result_acosh_ranges_dec128_t_is_ok      = local::test_acosh_ranges<decimal128_t>     (8);
+    const bool result_acosh_ranges_dec_fast32_t_is_ok  = local::test_acosh_ranges<decimal_fast32_t> (8);
+    const bool result_acosh_ranges_dec_fast64_t_is_ok  = local::test_acosh_ranges<decimal_fast64_t> (8);
+    const bool result_acosh_ranges_dec_fast128_t_is_ok = local::test_acosh_ranges<decimal_fast128_t>(8);
 
     BOOST_TEST(result_acosh_ranges_dec32_t_is_ok);
     BOOST_TEST(result_acosh_ranges_dec64_t_is_ok);
