### 2026-09-09 1eef246
Fix sign dropped by MSVC 14.2 Optimizer
