+++ include/boost/decimal/detail/int128/numeric.hpp
@@ -6,11 +6,14 @@
 #define BOOST_DECIMAL_DETAIL_INT128_NUMERIC_HPP
 
 #include <boost/decimal/detail/int128/bit.hpp>
+#include <boost/decimal/detail/int128/cstdlib.hpp>
 #include <boost/decimal/detail/int128/detail/traits.hpp>
+#include <boost/decimal/detail/int128/utilities.hpp>
 
 #ifndef BOOST_DECIMAL_DETAIL_INT128_BUILD_MODULE
 
 #include <limits>
+#include <utility>
 
 #endif
 
@@ -32,8 +35,8 @@ struct reduced_integers
                                  std::is_same<IntegerType, unsigned long>::value ||
                                  std::is_same<IntegerType, signed long long>::value ||
                                  std::is_same<IntegerType, unsigned long long>::value ||
-                                 std::is_same<IntegerType, int128_t>::value ||
-                                 std::is_same<IntegerType, uint128_t>::value};
+                                 std::is_same<IntegerType, int128>::value ||
+                                 std::is_same<IntegerType, uint128>::value};
 };
 
 #if defined(BOOST_DECIMAL_DETAIL_INT128_HAS_INT128) || defined(BOOST_DECIMAL_DETAIL_INT128_HAS_MSVC_INT128)
@@ -52,25 +55,25 @@ BOOST_DECIMAL_DETAIL_INT128_INLINE_CONSTEXPR bool is_reduced_integer_v {reduced_
 
 } // namespace detail
 
-BOOST_DECIMAL_DETAIL_INT128_EXPORT BOOST_DECIMAL_DETAIL_INT128_HOST_DEVICE constexpr uint128_t add_sat(const uint128_t x, const uint128_t y) noexcept
+BOOST_DECIMAL_DETAIL_INT128_EXPORT BOOST_DECIMAL_DETAIL_INT128_HOST_DEVICE constexpr uint128 saturating_add(const uint128 x, const uint128 y) noexcept
 {
     const auto z {x + y};
 
     if (z < x)
     {
-        return (std::numeric_limits<uint128_t>::max)();
+        return (std::numeric_limits<uint128>::max)();
     }
 
     return z;
 }
 
-BOOST_DECIMAL_DETAIL_INT128_EXPORT BOOST_DECIMAL_DETAIL_INT128_HOST_DEVICE constexpr uint128_t sub_sat(const uint128_t x, const uint128_t y) noexcept
+BOOST_DECIMAL_DETAIL_INT128_EXPORT BOOST_DECIMAL_DETAIL_INT128_HOST_DEVICE constexpr uint128 saturating_sub(const uint128 x, const uint128 y) noexcept
 {
     const auto z {x - y};
 
     if (z > x)
     {
-        return (std::numeric_limits<uint128_t>::min)();
+        return (std::numeric_limits<uint128>::min)();
     }
 
     return z;
@@ -82,50 +85,50 @@ BOOST_DECIMAL_DETAIL_INT128_EXPORT BOOST_DECIMAL_DETAIL_INT128_HOST_DEVICE const
 #  pragma warning(disable : 4146) // Unary minus applied to unsigned type
 #endif
 
-BOOST_DECIMAL_DETAIL_INT128_HOST_DEVICE constexpr int128_t add_sat(const int128_t x, const int128_t y) noexcept
+BOOST_DECIMAL_DETAIL_INT128_EXPORT BOOST_DECIMAL_DETAIL_INT128_HOST_DEVICE constexpr int128 saturating_add(const int128 x, const int128 y) noexcept
 {
     // Detect overflow BEFORE the addition to avoid signed overflow UB.
[…]
