+++ test/benchmarks.cpp
@@ -3,9 +3,6 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 
-#define BOOST_DECIMAL_DETAIL_INT128_ALLOW_SIGN_CONVERSION
-#define BOOST_DECIMAL_DETAIL_INT128_ALLOW_SIGN_COMPARE
-
 #include <boost/decimal.hpp>
 #include <chrono>
 #include <random>
