### 2026-09-09 a7be29a
Move specializations out of int128 and into our own lib
