### 2026-09-09 f0f12a1
Make the remainder of a u256 divide by one word correct (#1448)
- The single word branch of impl::div_mod wrote the remainder in word 0
of the array of
the dividend. The words above word 0 kept the dividend, thus from_words
gave the
  dividend back with its lowest 32 bits replaced.
- One 32 bit word holds the divisors 10^1 to 10^9. From 10^10 the
divisor needs two
words, thus the Knuth path runs, and that path writes the full
remainder.
- The rounding reads the remainder to learn whether a digit which is not
zero went away.
  A wrong remainder moves an exact result one step away from zero.
- T
[…]
