### 2026-09-11 5814468
Fix shadowing warning
