### 2026-09-12 ad4de63
Round a sum of far apart operands by the sign of the result
- The far apart branch of add_impl rounded the magnitude of the dominant operand and not
  the value of the sum. Upward on a negative result went away from zero, and downward on a
  negative result went toward zero.
- The upward branch also took one from a positive result which shrinks, and its copy for
  the right operand had no check for a power of ten.
- One block now selects the dominant operand, reads the sign of the result, and moves the
  magnitude away from zero or toward zero as the mode asks.
- Four expectations of test/gith
[…]
