### 2026-09-12 428f8f0
Ignore padding warning on MSVC for unused facility
