### 2026-09-12 bad17ca
Round a sum of far apart operands by the sign of the result (#1452)
