### 2026-09-13 34b1d05
Round a quotient of a non-default mode by its remainder
- The divides of the non-default rounding modes dropped the remainder of the wide quotient
  before the constructor rounded it. A mode which must move away from the truncated
  quotient could not.
- The 32 and 64 bit types held five or six digits past the precision, thus the defect
  showed only when all of them were zero. The 128 bit types held zero or one digit, thus
  most quotients were wrong, and fe_dec_to_nearest_from_zero was wrong too.
- The remainder now goes in as one sticky digit after the wide quotient. The 128 bit
  divide al
[…]
