### 2026-09-14 9074892
Give an exact cancel the sign of its rounding mode
- IEEE 754-2019 6.3: a sum which cancels exactly is +0 in every mode but downward,
  where it is -0.
- The aligned kernel took the sign of its left operand, thus -1 + 1 was -0 in the
  default mode.
- The other paths gave +0 in every mode, thus 1 - 1 was +0 in the downward mode.
- The kernel now gives +0, and a new helper gives the sign of the mode to both add
  paths and to the 128 bit fma.
- fesetround did not read its argument without constant evaluation detection, thus
  MSVC C++14 gave C4100 for the new test in that configuration.

Fixes 
[…]
