### 2026-09-15 ed63228
Give an exact cancel the sign of its rounding mode (#1456)
