### 2026-09-16 1c8cdd8
Correct the spelling of rounding_mode under the toward zero macro (#1458)
