+++ include/boost/decimal/cfenv.hpp
@@ -72,7 +72,7 @@ BOOST_DECIMAL_INLINE_CONSTEXPR_VARIABLE auto _boost_decimal_global_rounding_mode
     #elif defined(BOOST_DECIMAL_FE_DEC_TO_NEAREST_FROM_ZERO)
     rounding_mode::fe_dec_to_nearest_from_zero
     #elif defined (BOOST_DECIMAL_FE_DEC_TOWARD_ZERO)
-    rounding_mde::fe_dec_toward_zero
+    rounding_mode::fe_dec_toward_zero
     #elif defined(BOOST_DECIMAL_FE_DEC_UPWARD)
     rounding_mode::fe_dec_upward
     #else
