+++ include/boost/decimal/decimal_fast32_t.hpp
@@ -562,7 +562,12 @@ constexpr decimal_fast32_t::decimal_fast32_t(T1 coeff, T2 exp, const detail::con
     const auto biased_exp {static_cast<int>(exp) + detail::bias};
 
     // decimal32_t exponent holds 8 bits
-    if (biased_exp > detail::max_biased_exp_v<decimal_fast32_t>)
+    if (biased_exp > detail::max_biased_exp_v<decimal_fast32_t> && detail::overflow_is_finite(is_negative))
+    {
+        significand_ = static_cast<significand_type>(max_normal_significand);
+        exponent_ = static_cast<exponent_type>(detail::max_biased_exp_v<decimal_fast32_t>);
+    }
+    else if (biased_exp > detail::max_biased_exp_v<decimal_fast32_t>)
     {
         significand_ = detail::d32_fast_inf;
     }
