### 2026-09-19 628546d
Give an overflow the largest finite value in the modes which ask for it (#1460)
- IEEE 754-2019 7.4: an overflow is the largest finite value in the
toward zero mode,
and in the directed mode which points at zero for the sign. Every type
gave an
  infinity in every mode.
- A new helper reads the mode, and each constructor packs the largest
finite value in
  place of the infinity when the helper says so.
- The float constructors of the three IEEE types had their own overflow
check, which
  lost the sign. They now take the same path as the fast types.

Fixes #1459
