### 2026-09-20 6499302
Round sqrt correctly in the current rounding mode
- IEEE 754-2019 4.3 and 5.4.1: sqrt gives the infinitely precise root
  rounded in the current mode. The 32 and 64 bit kernels truncated the
  integer root, an odd exponent multiplied the root by a rounded sqrt(10)
  for the three sizes, and no kernel read the mode.
- For an odd exponent the three kernels take the root of 10 * gx, so the
  result is rounded once. The estimate is scaled by an integer sqrt(10),
  and the Newton steps correct it. The multiply by sqrt10_v goes away.
- A new helper reads the mode, and the final step of each kernel k
[…]
