+++ test/test_sqrt_rounding.cpp
@@ -1,0 +1,167 @@
+// Copyright 2026 Matt Borland
+// Distributed under the Boost Software License, Version 1.0.
+// https://www.boost.org/LICENSE_1_0.txt
+//
+// IEEE 754-2019 4.3 and 5.4.1: sqrt is correctly rounded in the current mode. The 32 and 64
+// bit kernels truncated the root, an odd exponent multiplied by a rounded sqrt(10), and no
+// kernel read the mode.
+
+#include <boost/decimal.hpp>
+#include <boost/core/lightweight_test.hpp>
+#include <iostream>
+
+using namespace boost::decimal;
+using namespace boost::decimal::literals;
+
+template <typename T> struct fast_of;
+template <> struct fast_of<decimal32_t> { using type = decimal_fast32_t; };
+template <> struct fast_of<decimal64_t> { using type = decimal_fast64_t; };
+template <> struct fast_of<decimal128_t> { using type = decimal_fast128_t; };
+
+// The value of the mode from the floor and the ceiling of the root. A root is never a tie.
+template <typename T>
+constexpr auto expected(const rounding_mode mode, const T floor, const T ceil, const bool near_up) -> T
+{
+    return mode == rounding_mode::fe_dec_upward ? ceil :
+           mode == rounding_mode::fe_dec_toward_zero || mode == rounding_mode::fe_dec_downward ? floor :
+           near_up ? ceil : floor;
+}
+
+template <typename T>
+void check_mode(const T x, const rounding_mode mode, const T value)
+{
+    using F = typename fast_of<T>::type;
+    fesetround(mode);
+    BOOST_TEST_EQ(sqrt(x), value);
+    BOOST_TEST_EQ(sqrt(static_cast<F>(x)), static_cast<F>(value));
+}
+
+// floor and ceil come from an integer square root at the precision of the type
+template <typename T>
+void check(const T x, const T floor, const T ceil, const bool near_up)
+{
+    #ifdef BOOST_DECIMAL_NO_CONSTEVAL_DETECTION
+    // fesetround has no effect here, thus only the compile-time mode is tested
+    check_mode(x, _boost_decimal_global_rounding_mode, expected(_boost_decimal_global_rounding_mode, floor, ceil, near_up));
+    #else
+    check_mode(x, rounding_mode::fe_dec_to_nearest, expected(rounding_mode::fe_dec_to_nearest, floor, ceil, near_up));
+    check_mode(x, rounding_mode::fe_dec_to_nearest_from_zero, expected(rounding_mode::fe_dec_to_nearest_from_zero, floor, ceil, near_up));
+    check_mode(x, rounding_mode::fe_dec_toward_zero, floor);
+    check_mode(x, rounding_mode::fe_dec_downward, floor);
+    check_mode(x, rounding_mode::fe_dec_upward, ceil);
+    #endif
+}
+
+template <typename T>
+void check(const int n, const T floor, const T ceil, const bool near_up)
+{
+    check(T {n, 0}, floor, ceil, near_up);
+}
+
+int main()
+{
+    constexpr auto mode {_boost_decimal_global_rounding_mode};
+    static_assert(sqrt(decimal32_t {11, 0}) == expected(mode, decimal32_t {3316624, -6}, decimal32_t {3316625, -6}, true), "");
+    static_assert(sqrt(decimal64_t {26, 0}) == expected(mode, decimal64_t {UINT64_C(5099019513592784), -15}, decimal64_t {UINT64_C(5099019513592785), -15}, true), "");
+
[…]
