### 2026-09-22 1269876
Name the file of the to_from_file example after its executable
- b2 runs the variants of the example in one directory, some of them at the same time,
  and they all used example_values.txt. One variant then read a file which another one
  had removed, and exited with 1.
- The name now comes from argv[0], so every variant has a file of its own. Two copies of
  the old example fail 4 times in 100 parallel runs, the new one never.
- The standard allows argc to be 0, so the name falls back to a literal in that case.
- The write side opens the file in binary mode like the read side, which matters o
[…]
