+++ .github/workflows/ci.yml
@@ -34,34 +34,6 @@ jobs:
       fail-fast: false
       matrix:
         include:
-          - toolset: gcc-8
-            cxxstd: "03,11,14,17,2a"
-            address_model: 32,64
-            os: ubuntu-latest
-            container: ubuntu:18.04
-            install:
-              - g++-8-multilib
-          - toolset: gcc-9
-            cxxstd: "03,11,14,17,2a"
-            address_model: 32,64
-            os: ubuntu-latest
-            container: ubuntu:20.04
-            install:
-              - g++-9-multilib
-          - toolset: gcc-9
-            cxxstd: "03-gnu,11-gnu,14-gnu,17-gnu,2a-gnu"
-            address_model: 32,64
-            os: ubuntu-latest
-            container: ubuntu:20.04
-            install:
-              - g++-9-multilib
-          - toolset: gcc-10
-            cxxstd: "03,11,14,17,20"
-            address_model: 32,64
-            os: ubuntu-latest
-            container: ubuntu:20.04
-            install:
-              - g++-10-multilib
           - toolset: gcc-11
             cxxstd: "03,11,14,17,20,23"
             address_model: 32
@@ -145,44 +117,6 @@ jobs:
               - g++-14-multilib
 
           # Linux, clang
-          - toolset: clang
-            compiler: clang++-6.0
-            cxxstd: "03,11,14,17"
-            os: ubuntu-latest
-            container: ubuntu:18.04
-            install:
-              - clang-6.0
-          - toolset: clang
-            compiler: clang++-7
-            cxxstd: "03,11,14,17"
-            os: ubuntu-latest
-            container: ubuntu:18.04
-            install:
-              - clang-7
-          # Note: clang-8 does not fully support C++20, so it is not compatible with libstdc++-8 in this mode
-          - toolset: clang
-            compiler: clang++-8
-            cxxstd: "03,11,14,17,2a"
-            os: ubuntu-latest
-            container: ubuntu:18.04
-            install:
-              - clang-8
-              - g++-7
-            gcc_toolchain: 7
-          - toolset: clang
-            compiler: clang++-9
-            cxxstd: "03,11,14,17,2a"
-            os: ubuntu-latest
-            container: ubuntu:20.04
-            install:
-              - clang-9
-          - toolset: clang
-            compiler: clang++-10
-            cxxstd: "03,11,14,17,20"
-            os: ubuntu-latest
-            container: ubuntu:20.04
-            install:
-              - clang-10
           - toolset: clang
             compiler: clang++-11
             cxxstd: "03,11,14,17"
@@ -350,7 +284,7 @@ jobs:
                 fi
             fi
             git config --global pack.threads 0
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
 
       - name: Install packages
         if: matrix.install
@@ -559,7 +493,7 @@ jobs:
     runs-on: ${{matrix.os}}
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
 
       - name: Setup Boost
         shell: cmd
@@ -601,7 +535,7 @@ jobs:
     runs-on: ${{matrix.os}}
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
 
       - name: Install packages
         if: matrix.install
@@ -648,7 +582,7 @@ jobs:
     runs-on: ${{matrix.os}}
 
     steps:
[…]
