### 2026-09-23 c1baf9d
Classify NaN and the infinities without a pragma or an exception
- The float constructors of the six types tested NaN with != and ±inf with == against
  numeric_limits under a -Wfloat-equal pragma. GCC 8 to 14 do not apply the pragmas of a
  precompiled header to the templates which they instantiate in the including file, so the
  pch commit needs that region gone.
- A NaN now comes from __builtin_isnan, which raises nothing. An ordered compare raises
  FE_INVALID for a NaN, and == or != draws the warning. MSVC has no such warning, so it
  keeps the compare with !=.
- An infinity now comes fro
[…]
