### 2026-09-23 97be1eb
Pick the fast type of from_bid by an overload instead of a specialization
- Clang 15 to 18 drop an explicit specialization of a constrained function template when
  it comes from a precompiled header, and the from_bid<fast> calls then fail to link.
- The fast type now has its own overload, constrained with is_fast_type_v, and the
  primary is constrained with is_ieee_type_v, so exactly one overload matches.
- The fast overload asks for is_decimal_floating_point_v as well, because is_fast_type_v
  is the negation of is_ieee_type_v and is true for int, double and every other type.
- test_bid_con
[…]
