### 2026-09-23 fb7d32c
Drop the Node 20 workaround with the containers which needed it
- Only the Ubuntu 16 and 18 containers could not run the Node 20 of actions/checkout, and
  the matrix has none of them any more.
- The mount of /node20217 and the download into it are therefore dead, so both go.
- ACTIONS_ALLOW_USE_UNSECURE_NODE_VERSION goes too, because no action of ci.yml needs an
  old Node any more.
