### 2026-09-24 967bd7d
Compute log and log10 near 1 with log1p
- Above 1, log took log10(x) * ln10, and the reduction of log10 subtracts terms near 1.
  Below 1, both took the value of 1/x, which rounds. Near 1 the two lose most digits,
  and log(0.9999999_DF) gives -0.
- log2 calls log, so it lost the same digits.
- For x in [0.5, 1.5], x - 1 is exact, and log1p takes its series branch there, which
  does not call log. log and log10 now take log1p(x - 1) in this range.
- log keeps x equal to 1 out of the new branch, so the downward mode still gives +0.
- github_issue_1110 expected the old log10 value 4.4e-33, and n
[…]
