### 2026-09-24 fb2360f
Round log10 and log2 near 1 only once
- log10 divided the rounded log1p by ln10, and log2 divided the rounded log by ln2. Near 1
  these two more roundings gave up to 4 units in the last digit.
- log1p_parts gives the two parts of the series of log1p before its last rounding, and
  log1p itself uses it as before.
- log10 and log2 now scale the two parts with 2/ln(10) and 2/ln(2) in two parts each, and
  round only once.
- The series of log1p used all terms of its table for each argument, thus log, log10 and
  log2 near 1 became slower than the old path. log1p_series_sum now uses only the terms
[…]
