+++ include/boost/decimal/detail/cmath/impl/log1p_impl.hpp
@@ -9,11 +9,15 @@
 #include <boost/decimal/fwd.hpp>
 #include <boost/decimal/detail/concepts.hpp>
 #include <boost/decimal/detail/cmath/impl/taylor_series_result.hpp>
+#include <boost/decimal/detail/int128.hpp>
+#include <boost/decimal/detail/power_tables.hpp>
+#include <boost/decimal/detail/promotion.hpp>
 
 #ifndef BOOST_DECIMAL_BUILD_MODULE
 #include <array>
 #include <cstddef>
 #include <cstdint>
+#include <type_traits>
 #endif
 
 namespace boost {
@@ -200,6 +204,36 @@ constexpr typename log1p_table_imp<b>::d128_fast_coeffs_t log1p_table_imp<b>::d1
 
 using log1p_table = log1p_detail::log1p_table_imp<true>;
 
+// Horner over the first n terms only. With -log10(z2) >= L, (n + 1) * L >= digits + 2
+// keeps each dropped term less than 0.01 of the last digit of the result.
+template <typename T, typename Array>
+constexpr auto log1p_series_sum(const T z2, const Array& coeffs) noexcept -> T
+{
+    constexpr int digits {std::numeric_limits<T>::digits10};
+    // ceil(1000 * log10(d + 1)) for the leading digit d of z2
+    constexpr int log10_next[] {0, 302, 478, 603, 699, 779, 846, 904, 955, 1000};
+
+    int exp10 {};
+    const auto sig {frexp10(z2, &exp10)};
+    const int lead {static_cast<int>(sig / detail::pow10(static_cast<decltype(sig)>(digits - 1)))};
+    const int bound {1000 * (1 - digits - exp10) - log10_next[lead]};  // 1000 * L
+
+    std::size_t n {coeffs.size()};
+    if (bound > 0)
+    {
+        const auto need {static_cast<std::size_t>((1000 * (digits + 2) + bound - 1) / bound - 1)};
+        n = need < 1U ? 1U : (need < n ? need : n);
+    }
+
+    auto result {coeffs[n - 1U]};
+    for (std::size_t i {n - 1U}; i-- > 0U;)
+    {
+        result = unchecked_fma(result, z2, coeffs[i]);
+    }
+
+    return result;
+}
+
 // 2*atanh(w) = 2*w + (2/3)*w^3 + (2/5)*w^5 + ...
 // The first coefficient is exactly 2 for every type, thus the tables leave it out and
 // the caller adds that term itself. This gives the rest of the series, divided by w^3.
@@ -209,37 +243,84 @@ constexpr auto log1p_series_tail(T z2) noexcept;
 template <>
 constexpr auto log1p_series_tail<decimal32_t>(decimal32_t z2) noexcept
 {
-    return taylor_series_result(z2, log1p_table::d32_coeffs);
+    return log1p_series_sum(z2, log1p_table::d32_coeffs);
 }
 
 template <>
 constexpr auto log1p_series_tail<decimal_fast32_t>(decimal_fast32_t z2) noexcept
 {
-    return taylor_series_result(z2, log1p_table::d32_fast_coeffs);
+    return log1p_series_sum(z2, log1p_table::d32_fast_coeffs);
 }
 
 template <>
 constexpr auto log1p_series_tail<decimal64_t>(decimal64_t z2) noexcept
 {
-    return taylor_series_result(z2, log1p_table::d64_coeffs);
+    return log1p_series_sum(z2, log1p_table::d64_coeffs);
 }
 
 template <>
 constexpr auto log1p_series_tail<decimal_fast64_t>(decimal_fast64_t z2) noexcept
 {
[…]
