### 2026-09-24 2a3bc58
Use the optimized multiply path for decimal32_t
- decimal32_t * decimal32_t gave the raw product to the constructor, and did not use the
  faster path of the other types.
- That path rounds to 7 digits before pack_in_range, and the constructor then rounds a
  subnormal result again. Thus a product with an exponent below etiny still goes to the
  constructor unrounded. Above etiny, the rounded result has 7 digits or is at least 1e-94,
  so one rounding is correct.
- decimal32_t * Integer keeps the constructor path, because the optimized path gives a
  zero product a different exponent.
- The re
[…]
