### 2026-09-25 271a8bb
Divide the decimal128 product by 10^33 or 10^34 with a 128-bit reciprocal (#1466)
- `mul_finalize_u256` gave the 67 or 68 digit product to
  `divmod_pow10_u256`. That function does a 512-bit multiply-high, a
  256-bit multiply and a shift loop. A profile showed about 64% of a
  `decimal128_t` multiply in it.
- 10^k is 5^k * 2^k, with k = 33 or 34. The product shifted right by k
bits
  is less than 2^192. A 128-bit reciprocal of 5^k gives the quotient of
  this value or one less, and one compare corrects it.
- The remainder is less than 2 * 5^k, which is less than 2^80, so its
low
128 bits are 
[…]
