### 2026-09-25 ea218b6
Use the optimized multiply path for decimal32_t (#1465)
- `decimal32_t * decimal32_t` gave the raw product to the constructor.
The
  other types take the path through `mul_finalize_u64`, which is faster.
  A comment said that the hardware divide of the constructor was cheap
  enough, but on the machine below the constructor path is slower.
- `decimal32_t` now takes the same path as the other types, except for
one
  case. That path rounds the product to 7 digits before `pack_in_range`,
  and the constructor then rounds a subnormal result a second time.
- Because of this, a product with an expon
[…]
