### 2026-09-25 f694b0e
Keep the digits of log, log10 and log2 near 1 (#1468)
