+++ include/boost/decimal/detail/cmath/log1p.hpp
@@ -25,6 +25,41 @@ namespace decimal {
 
 namespace detail {
 
+// log1p(x) is 2*wh + s for |x| not more than 1/2, before the last rounding. log10 and
+// log2 scale the two parts near 1, so that they round only once.
+template <typename T>
+struct log1p_parts_t
+{
+    T wh;
+    T s;
+};
+
+template <typename T>
+constexpr auto log1p_parts(const T x) noexcept -> log1p_parts_t<T>
+{
+    // log1p(x) = 2 * atanh(w), with w = x / (2 + x). For |x| not more than
+    // 1/2 the value of |w| is not more than 1/3, which the coefficient table
+    // covers. The first branch takes every larger |x| through log(x + 1).
+    // Two corrections make the result increase at every argument. The sum
+    // 2 + x is not exact, and its error makes the quotient fall where x rises,
+    // thus e holds that error and wh + wl holds the reduction to about two
+    // times the digits of the type. The first coefficient is exactly 2, thus
+    // 2*wh stays out of the rounding of the series and only tail rounds.
+    constexpr T two { 2, 0 };
+
+    const T d  { two + x };
+    const T e  { x - (d - two) };  // 2 + x == d + e, exact
+    const T wh { x / d };
+    const T r  { detail::unchecked_fma(-wh, d, x) - wh * e };
+    const T wl { r / d };
+
+    const T w    { wh + wl };
+    const T y    { w * w };
+    const T tail { w * y * detail::log1p_series_tail(y) };
+
+    return { wh, detail::unchecked_fma(two, wl, tail) };
+}
+
 template <typename T>
 constexpr auto log1p_impl(const T x) noexcept
     BOOST_DECIMAL_REQUIRES(detail::is_decimal_floating_point_v, T)
@@ -75,27 +110,11 @@ constexpr auto log1p_impl(const T x) noexcept
         }
         else
         {
-            // log1p(x) = 2 * atanh(w), with w = x / (2 + x). For |x| not more than
-            // 1/2 the value of |w| is not more than 1/3, which the coefficient table
-            // covers. The first branch takes every larger |x| through log(x + 1).
-            // Two corrections make the result increase at every argument. The sum
-            // 2 + x is not exact, and its error makes the quotient fall where x rises,
-            // thus e holds that error and wh + wl holds the reduction to about two
-            // times the digits of the type. The first coefficient is exactly 2, thus
-            // 2*wh stays out of the rounding of the series and only tail rounds.
             constexpr T two { 2, 0 };
 
-            const T d  { two + x };
-            const T e  { x - (d - two) };  // 2 + x == d + e, exact
-            const T wh { x / d };
-            const T r  { detail::unchecked_fma(-wh, d, x) - wh * e };
-            const T wl { r / d };
+            const auto parts { detail::log1p_parts(x) };
 
-            const T w    { wh + wl };
-            const T y    { w * w };
-            const T tail { w * y * detail::log1p_series_tail(y) };
-
-            result = detail::unchecked_fma(two, wh, detail::unchecked_fma(two, wl, tail));
+            result = detail::unchecked_fma(two, parts.wh, parts.s);
         }
     }
 
