+++ test/github_issue_1467.cpp
@@ -1,0 +1,80 @@
+// Copyright 2026 Shen-Ta Hsieh
+// Distributed under the Boost Software License, Version 1.0.
+// https://www.boost.org/LICENSE_1_0.txt
+
+#include <boost/decimal.hpp>
+#include <boost/core/lightweight_test.hpp>
+#include <limits>
+
+using namespace boost::decimal;
+using namespace boost::decimal::literals;
+
+// The expected values are the MPFR values rounded to the type. Near 1 the old code
+// lost most digits, so a relative error of one epsilon still separates the two.
+template <typename T>
+void check(const T x, const T expected_log, const T expected_log10, const T expected_log2)
+{
+    constexpr T eps {std::numeric_limits<T>::epsilon()};
+    BOOST_TEST_LE(abs(log(x) - expected_log), eps * abs(expected_log));
+    BOOST_TEST_LE(abs(log10(x) - expected_log10), eps * abs(expected_log10));
+    BOOST_TEST_LE(abs(log2(x) - expected_log2), eps * abs(expected_log2));
+}
+
+int main()
+{
+    check(1.000001_DF, 9.999995e-7_DF, 4.342943e-7_DF, 1.442694e-6_DF);
+    check(0.9999999_DF, -1.000000e-7_DF, -4.342945e-8_DF, -1.442695e-7_DF);
+    check(1.000123_DF, 1.229924e-4_DF, 5.341494e-5_DF, 1.774406e-4_DF);
+    check(0.999377_DF, -6.231941e-4_DF, -2.706498e-4_DF, -8.990791e-4_DF);
+    check(1.049999_DF, 4.878921e-2_DF, 2.118889e-2_DF, 7.038795e-2_DF);
+    check(0.9500001_DF, -5.129319e-2_DF, -2.227635e-2_DF, -7.400043e-2_DF);
+    check(1.499999_DF, 4.054644e-1_DF, 1.760910e-1_DF, 5.849615e-1_DF);
+    check(0.5000001_DF, -6.931470e-1_DF, -3.010299e-1_DF, -9.999997e-1_DF);
+
+    check(1.000001_DFF, 9.999995e-7_DFF, 4.342943e-7_DFF, 1.442694e-6_DFF);
+    check(0.9999999_DFF, -1.000000e-7_DFF, -4.342945e-8_DFF, -1.442695e-7_DFF);
+    check(1.000123_DFF, 1.229924e-4_DFF, 5.341494e-5_DFF, 1.774406e-4_DFF);
+    check(0.999377_DFF, -6.231941e-4_DFF, -2.706498e-4_DFF, -8.990791e-4_DFF);
+    check(1.049999_DFF, 4.878921e-2_DFF, 2.118889e-2_DFF, 7.038795e-2_DFF);
+    check(0.9500001_DFF, -5.129319e-2_DFF, -2.227635e-2_DFF, -7.400043e-2_DFF);
+    check(1.499999_DFF, 4.054644e-1_DFF, 1.760910e-1_DFF, 5.849615e-1_DFF);
+    check(0.5000001_DFF, -6.931470e-1_DFF, -3.010299e-1_DFF, -9.999997e-1_DFF);
+
+    check(1.000000000000001_DD, 9.999999999999995e-16_DD, 4.342944819032516e-16_DD, 1.442695040888963e-15_DD);
+    check(0.9999999999999999_DD, -1.000000000000000e-16_DD, -4.342944819032518e-17_DD, -1.442695040888963e-16_DD);
+    check(1.000123_DD, 1.229924361202318e-4_DD, 5.341493632285486e-5_DD, 1.774405776575110e-4_DD);
[…]
