+++ include/boost/decimal/detail/cmath/log2.hpp
@@ -12,6 +12,7 @@
 #include <boost/decimal/detail/config.hpp>
 #include <boost/decimal/numbers.hpp>
 #include <boost/decimal/detail/cmath/log.hpp>
+#include <boost/decimal/detail/cmath/log1p.hpp>
 
 #ifndef BOOST_DECIMAL_BUILD_MODULE
 #include <cmath>
@@ -27,6 +28,17 @@ template <typename T>
 constexpr auto log2_impl(const T x) noexcept
     BOOST_DECIMAL_REQUIRES(detail::is_decimal_floating_point_v, T)
 {
+    constexpr T one { 1 };
+
+    if ((x != one) && (x >= T { 5, -1 }) && (x <= T { 15, -1 }))
+    {
+        // As in log10, the two parts of log1p near 1 round only once.
+        const auto parts { detail::log1p_parts(x - one) };
+        constexpr auto k { detail::two_over_ln2<T>() };
+
+        return detail::unchecked_fma(parts.wh, k.hi, detail::unchecked_fma(parts.wh, k.lo, parts.s * numbers::log2e_v<T>));
+    }
+
     return ::boost::decimal::log(x) / numbers::ln2_v<T>;
 }
 
