### 2026-09-26 31ab53d
Add and tighten tols log tests
