### 2026-09-26 79c2580
Handle a new compiler warning
